### 2017-12-20 #175 @elinw
Skimr
Creates compact and flexible data summaries that are pipeable and display nicely in the console. For a given data frame or vector the skim function provides a useful set of summary statistics (based on the class of the individual vector/column)  that allows users to skim their data to get an overall sense of what is included, extent of missing values, and similar information.  The skim generic can be extended by users to other data structures.

```
 Package: skimr
Title: Compact and Flexible Summaries of Data
Version: 0.9.92
  person("Amelia", "McNamara", email="amcnamara@smith.edu", role = "aut"),
  person("Eduardo", "Arino de la Rubia", email="earino@gmail.com", role = "aut"),
  person("Hao", "Zhu", email="haozhu233@gmail.com", role = "aut"),
  person("Julia", "Lowndes", email="lowndes@nceas.ucsb.edu", role = 'ctb'),
  person("Shannon", "Ellis", email="sellis18@jhmi.edu", role = "aut"),
  person("Elin", "Waring", email="elin.waring@gmail.com", role = "cre"),
  person("Michael", "Quinn", email="msquinn2@illinois.edu", role = "aut"),
  person("Hope", "McLeod", email="hmgit2@gmail.com", role = 'ctb'),
  person("Hadley", "Wickham", email="hadley@rstudio.com", role = 'ctb'),
  person("Connor", "Kirkpatrick", email="hello@connorkirkpatrick.com", role = 'ctb')
  )
Maintainer: Elin Waring <elin.waring@gmail.com>
Description: A simple to use summary function that can be used with pipes
    and displays nicely in the console. The default summary statistics may be
    modified by the user as can the default formatting. Support for data frames
    and vectors is included, and users can implement their own skim methods for
    specific object types as described in a vignette. Default summaries include
    support for inline spark graphs. Instructions for managing these on
    specific operating systems are given in the Using skimr vignette and the
    README.
    dplyr (>= 0.7),
    pander,
    tibble (>= 0.6),
    tidyr (>= 0.7),
    tidyselect
    extrafont,
    withr
License: GPL-3 + file LICENSE
Collate:
    'skimr-package.R'
    'formats.R'
    'stats.R'
    'functions.R'
    'options.R'
    'skim.R'
    'skim_print.R'
    'skim_v.R'
    'summary.R'

```

https://github.com/ropenscilabs/skimr

"tools"  Because skimr provides users with a way to get started with a new, unknown data set (by getting a quick overview (or skim) of the data that is readable and compact. It also serves as a good tool for reporting summary information about data.

* People who want to get a quick overview of a data set while initially reviewing, iteratively cleaning, or sharing with others.
* Beginning statistics students who  need simple access to a set of basic descriptive statistics.

This is intended as an improved version of the r core summary functions available. Like summary, skim is a generic that users can extend to any R data object.  It is designed to be somewhat like a more flexible version of fivenums() from `stats` or favstats() from the Mosaic package.

	- [ ] You intend to archive the code for the package in a long-term repository which meets the requirements of the journal.
