### 2017-12-21 b0583ac
Add missing contributors
