### 2017-12-21 d900afa
Add missing contributors
