### 2017-12-24 37d2abf
Clean up documentation.
