### 2017-12-29 2da11f3
Add minimum version for stringr.
