### 2018-01-02 c905d20
Remove documentation from internal spark functions.
