+++ R/stats.R
@@ -58,6 +58,7 @@ inline_hist <- function(x) {
   structure(spark_bar(hist_dt), class = c("spark", "character"))
 }
 
+
 #' Draw a sparkline bar graph with unicode block characters
 #'
 #' Rendered using [block elements](https://en.wikipedia.org/wiki/Block_Elements).
@@ -79,6 +80,8 @@ inline_hist <- function(x) {
 #'
 #' spark_bar(c(0, NA, 0.5, NA, 1))
 #' }
+#' @noRd
+
 spark_bar <- function(x, safe = TRUE) {
   stopifnot(is.numeric(x))
   
@@ -196,6 +199,7 @@ normalize01 <- function(x) {
   (x - min(x)) / (max(x) - min(x))
 }
 
+
 #' Draw a sparkline line graph with Braille characters.
 #'
 #' @inheritParams spark_bar
@@ -204,6 +208,8 @@ normalize01 <- function(x) {
 #' x <- seq(0, 1, length = 10)
 #' spark_line(x)
 #' }
+#' @noRd
+
 spark_line <- function(x) {
   stopifnot(is.numeric(x))
   
