### 2018-01-02 863cdc8
Force appropriate NA return from list_lengths_median.
