### 2018-01-02 788f265
Force converted matrix to use UTF-8 when printing.
