### 2018-01-04 eba796a
Fix issue with inline histograms when data are all NA.
