### 2018-01-05 fc7a3f9
Try enc2utf8 for printing the matrix.
