+++ NAMESPACE
@@ -35,6 +35,7 @@ export(n_missing)
 export(n_unique)
 export(num_range)
 export(one_of)
+export(pander)
 export(show_formats)
 export(show_skimmers)
 export(skim)
