### 2018-01-06 0fdba94
Move windows functions to utils file.
