### 2018-01-06 fbbee1f
Clean up tests.
