### 2018-01-07 64e1f5b
Put back deleted UTF fix.
