### 2018-01-07 bcff716
Add support for AppVeyor (windows CI)
