### 2018-01-07 4e2c377
Separate tests for 32-bit windows.
