+++ tests/testthat/test-skim_print.R
@@ -41,7 +41,8 @@ test_that("spark.print returns the correct result", {
   expect_identical(input, correct)
 })
 
-test_that("Skimr kable prints as expected", {
+test_that("Skimr kable prints as expected, 64-bit", {
+  skip_if(R.Version()$arch == "i386")
   skimmed <- skim(iris)
   inputRaw <- capture.output(skimr::kable(skimmed))
   input <- skimr:::fix_unicode(inputRaw)
@@ -88,6 +89,27 @@ test_that("Skimr kable prints as expected", {
   expect_equal(input[14], 
 "|:------------|:-------|:--------|:---|:----|:----|:---|:---|:------|:---|:---|:--------|"
   )
+})
+
+test_that("Skimr kable prints as expected, 32-bit windows", {
+  skip_if_not(R.Version()$arch == "i386")
+  skimmed <- skim(iris)
+  inputRaw <- capture.output(skimr::kable(skimmed))
+  input <- skimr:::fix_unicode(inputRaw)
+  
+  expect_length(input, 18)
+  expect_equal(input[15], 
+"|Petal.Length |0       |150      |150 |3.76 |1.77 |1   |1.6 |4.35   |5.1 |6.9 |▇▁▁▂▅▅▃▁ |"
+   )
+  expect_equal(input[16], 
+"|Petal.Width  |0       |150      |150 |1.2  |0.76 |0.1 |0.3 |1.3    |1.8 |2.5 |▇▁▁▃▃▃▂▂ |"
+  )
+  expect_equal(input[17], 
+"|Sepal.Length |0       |150      |150 |5.84 |0.83 |4.3 |5.1 |5.8    |6.4 |7.9 |▂▇▅▇▆▅▂▂ |"
+   )
+  expect_equal(input[18], 
+"|Sepal.Width  |0       |150      |150 |3.06 |0.44 |2   |2.8 |3      |3.3 |4.4 |▁▂▅▇▃▂▁▁ |"
+   )
 })
 
 test_that("skimr::pander prints as expected", {
