### 2018-01-07 47c1b35
Update news and exclude windows specific tests from Travis coverage.  Add test for make_utf8().
