### 2018-01-07 0c89598
One more line to exclude;  additional news.
