### 2018-01-08 fba7fcf
Update exclusions
