+++ .travis.yml
@@ -12,4 +12,4 @@ r_github_packages:
   - jimhester/covr
 
 after_success:
-  - Rscript -e 'covr::codecov(line_exclusions = list("R/skim_print.R" = c(147, 167, 169:171, 190)))'
+  - Rscript -e 'covr::codecov(line_exclusions = list("R/skim_print.R" = c(147, 167, 178:180, 190)))'
