### 2018-01-09 #175 @maelle
@elinw thanks for your submission & sorry about the delay! I've seen development is still active, do you consider `skimr` stable enough for review? It'd be best not to change too many things while the reviewers are looking.
