### 2018-01-09 9cd7411
Change max and min to p0 and p100 (use quantile for calculations) for numeric variables.
