+++ tests/testthat/test-functions.R
@@ -166,8 +166,8 @@ test_that("show_skimmers() lets you pick which type you want returned", {
 })
 
 test_that("show_skimmers() lets you pick which many types you want returned", {
-  correct <- list(numeric = c("missing", "complete", "n", "mean",  "sd", "min",
-                              "p25", "median", "p75", "max", "hist"),
+  correct <- list(numeric = c("missing", "complete", "n", "mean",  "sd", "p0",
+                              "p25", "median", "p75", "p100", "hist"),
                   character = c("missing",  "complete", "n", "min", "max",
                                 "empty", "n_unique"))
   skimmers <- show_skimmers(c("numeric", "character"))
@@ -188,8 +188,8 @@ test_that("show_skimmers() returns something if given an unassigned type", {
 
 test_that("Skim functions can be removed by setting them to NULL", {
   skim_with(numeric = list(hist = NULL))
-  correct <- list(numeric = c("missing", "complete", "n", "mean",  "sd", "min",
-                              "p25", "median", "p75", "max"))
+  correct <- list(numeric = c("missing", "complete", "n", "mean",  "sd", "p0",
+                              "p25", "median", "p75", "p100"))
   input <- show_skimmers("numeric")
   expect_identical(correct, input)
   skim_with_defaults()
