+++ tests/testthat/test-skim.R
@@ -17,7 +17,7 @@ test_that("Skimming a data frame works as expected", {
   expect_identical(input$type, c(rep("numeric", each = 11), 
                                  rep("factor", each = 12)))
   expect_identical(head(input$stat),
-                   c("missing", "complete", "n", "mean", "sd", "min"))
+                   c("missing", "complete", "n", "mean", "sd", "p0"))
   expect_identical(tail(input$stat), c(rep("top_counts", 5), rep("ordered", 1)))
   expect_identical(head(input$level), rep(".all", 6))
   expect_identical(tail(input$level),
@@ -67,7 +67,7 @@ test_that("Skimming a grouped data frame works as expected", {
                                     "carb"), 8, each = 11))
   expect_identical(input$type, rep("numeric", 792))
   expect_identical(input$stat, rep(c("missing", "complete", "n", "mean", "sd",
-                                     "min", "p25", "median", "p75", "max",
+                                     "p0", "p25", "median", "p75", "p100",
                                      "hist"), 72))
   expect_identical(input$level, rep(".all", 792))
   expect_identical(input$value[1:5], c(0, 1, 1, 21.5, NA))
@@ -95,7 +95,7 @@ test_that("skim_to_list works as expected", {
   expect_equal(dim(input[["numeric"]]), c(1, 12))
   expect_identical(names(input[["numeric"]]), 
                    c("variable", "missing", "complete", "n", "mean",
-                     "sd", "min", "p25", "median", "p75", "max", "hist" ))
+                     "sd", "p0", "p25", "median", "p75", "p100", "hist" ))
 })
 
 test_that("skim_to_list works with grouped data", {
@@ -137,7 +137,7 @@ test_that("Skimming a column of a data frame works as expected", {
   expect_identical(input$variable, c(rep(c("chickwts$weight"), each = 11)))
   expect_identical(input$type, c(rep("numeric", each = 11)))
   expect_identical(head(input$stat),
-                   c("missing", "complete", "n", "mean", "sd", "min"))
+                   c("missing", "complete", "n", "mean", "sd", "p0"))
   expect_identical(head(input$level), rep(".all", 6))
   expect_equal(head(input$value), c(0, 71, 71, 261.3, 78.1, 108), tol = .01)
   expect_identical(head(input$formatted),
@@ -164,7 +164,7 @@ test_that("Skimming a data frame with selected columns works as expected", {
   expect_identical(input$variable, rep("weight", 11))
   expect_identical(input$type, rep("numeric", 11))
   expect_identical(head(input$stat),
-                   c("missing", "complete", "n", "mean", "sd", "min"))
+                   c("missing", "complete", "n", "mean", "sd", "p0"))
   expect_identical(head(input$level), rep(".all", 6))
   expect_equal(head(input$value), c(0, 71, 71, 261.3, 78.1, 108), tol = .01)
   expect_identical(head(input$formatted),
@@ -214,7 +214,7 @@ test_that("Tidyselect helpers work as expected", {
                                        each = 11))
   expect_identical(input$type, rep("numeric", 22))
   expect_identical(head(input$stat),
-                   c("missing", "complete", "n", "mean", "sd", "min"))
+                   c("missing", "complete", "n", "mean", "sd", "p0"))
   expect_identical(head(input$level), rep(".all", 6))
   expect_equal(head(input$value), c(0, 150, 150, 5.84, 0.82, 4.3), 0.1)
   expect_identical(head(input$formatted),
@@ -242,7 +242,7 @@ test_that("Skimming a grouped df works as expected selecting two columns", {
   expect_identical(input$variable, rep(c("mpg", "disp"), 8, each = 11))
   expect_identical(input$type, rep("numeric", 176))
[…]
