+++ tests/testthat/test-skim_print.R
@@ -68,26 +68,26 @@ test_that("Skimr kable prints as expected, 64-bit", {
   expect_equal(input[11], "Variable type: numeric")
   expect_equal(input[12], "")  
   expect_equal(input[15], 
-"|Petal.Length |0       |150      |150 |3.76 |1.77 |1   |1.6 |4.35   |5.1 |6.9 |▇▁▁▂▅▅▃▁ |"
+"|Petal.Length |0       |150      |150 |3.76 |1.77 |1   |1.6 |4.35   |5.1 |6.9  |▇▁▁▂▅▅▃▁ |"
    )
   expect_equal(input[16], 
-"|Petal.Width  |0       |150      |150 |1.2  |0.76 |0.1 |0.3 |1.3    |1.8 |2.5 |▇▁▁▅▃▃▂▂ |"
+"|Petal.Width  |0       |150      |150 |1.2  |0.76 |0.1 |0.3 |1.3    |1.8 |2.5  |▇▁▁▅▃▃▂▂ |"
   )
   expect_equal(input[17], 
-"|Sepal.Length |0       |150      |150 |5.84 |0.83 |4.3 |5.1 |5.8    |6.4 |7.9 |▂▇▅▇▆▅▂▂ |"
+"|Sepal.Length |0       |150      |150 |5.84 |0.83 |4.3 |5.1 |5.8    |6.4 |7.9  |▂▇▅▇▆▅▂▂ |"
    )
   expect_equal(input[18], 
-"|Sepal.Width  |0       |150      |150 |3.06 |0.44 |2   |2.8 |3      |3.3 |4.4 |▁▂▅▇▃▂▁▁ |"
+"|Sepal.Width  |0       |150      |150 |3.06 |0.44 |2   |2.8 |3      |3.3 |4.4  |▁▂▅▇▃▂▁▁ |"
    )
   
   # The headers are different on windows
   # Just ignore them
   skip_on_os("windows")
   expect_equal(input[13], 
-"|variable     |missing |complete |n   |mean |sd   |min |p25 |median |p75 |max |hist     |"
+"|variable     |missing |complete |n   |mean |sd   |p0  |p25 |median |p75 |p100 |hist     |"
   )
   expect_equal(input[14], 
-"|:------------|:-------|:--------|:---|:----|:----|:---|:---|:------|:---|:---|:--------|"
+"|:------------|:-------|:--------|:---|:----|:----|:---|:---|:------|:---|:----|:--------|"
   )
 })
 
@@ -154,7 +154,7 @@ test_that("skimr::pander prints as expected", {
  "----------------------------------------------------------------------------"
   )
   expect_equal(input[23], 
- " variable   missing   complete   n     mean     sd     min    p25    median "
+ " variable   missing   complete   n     mean     sd     p0     p25    median "
   )
   expect_equal(input[24], 
  "---------- --------- ---------- ---- -------- ------- ----- ------- --------"
@@ -169,11 +169,11 @@ test_that("skimr::pander prints as expected", {
   expect_equal(input[28], "Table: Table continues below")
   expect_equal(input[29], "")
   expect_equal(input[30], " ")
-  expect_equal(input[31], "------------------------")
-  expect_equal(input[32], "  p75    max     hist   ")
-  expect_equal(input[33], "------- ----- ----------")
-  expect_equal(input[34], " 323.5   423   ▃▅▅▇▃▇▂▂ ")
-  expect_equal(input[35], "------------------------")
+  expect_equal(input[31], "-------------------------")
+  expect_equal(input[32], "  p75    p100     hist   ")
+  expect_equal(input[33], "------- ------ ----------")
[…]

[…]
