### 2018-01-09 #175 @maelle
I'm currently trying to perform editor checks. Tests fail which I think is due to encoding.

```r
> sessionInfo()
R version 3.4.0 Patched (2017-05-10 r72669)
Platform: x86_64-w64-mingw32/x64 (64-bit)
Running under: Windows >= 8 x64 (build 9200)

Matrix products: default

locale:
[1] LC_COLLATE=English_United States.1252  LC_CTYPE=English_United States.1252
[3] LC_MONETARY=English_United States.1252 LC_NUMERIC=C
[5] LC_TIME=English_United States.1252

attached base packages:

other attached packages:
[1] bindrcpp_0.2   testthat_1.0.2 skimr_1.0.1

 [1] Rcpp_0.12.14      rstudioapi_0.7    bindr_0.1         knitr_1.17
 [5] xml2_1.1.1        magrittr_1.5      roxygen2_6.0.1    devtools_1.13.2
 [9] tidyselect_0.2.3  R6_2.2.2          rlang_0.1.4       dplyr_0.7.4
[13] stringr_1.2.0     tools_3.4.0       withr_2.1.0       commonmark_1.2
[17] digest_0.6.12     assertthat_0.2.0  rprojroot_1.2     tibble_1.4.1
[21] crayon_1.3.4      tidyr_0.7.2       purrr_0.2.4       memoise_1.1.0
[25] glue_1.1.1        stringi_1.1.5     pillar_0.0.0.9000 compiler_3.4.0
[29] pander_0.6.1      desc_1.1.0        backports_1.1.1   pkgconfig_2.0.1
```

`testthat` output

````r
==> devtools::test()

Loading skimr
Loading required package: testthat
Testing skimr
Formatting options: ............W...
Change functions used by skim: ..........................................
Skim a data.frame: ................................................................................................................................
Print a skim_df object: ...............12SS3
Skim a vector within a data frame: 4..5..........6......
Use individual statistics functions on a vector: 789....ab.c...............
Make and print a summary of a skim object: .....................................

Skipped ------------------------------------------------------------------------
1. Skimr kable prints as expected, 32-bit windows (@test-skim_print.R#95) - R.Version()$arch == "i386"

2. skimr::pander prints as expected (@test-skim_print.R#117) - On windows

Warnings -----------------------------------------------------------------------
1. Formatting options change printed output (@test-formats.R#31) - package 'bindrcpp' was built under R version 3.4.1

Failed -------------------------------------------------------------------------
1. Failure: spark.print returns the correct result (@test-skim_print.R#41) -----
`input` not identical to `correct`.
1/1 mismatches
x[1]: "<U+2583><U+2585><U+2585><U+2587><U+2583><U+2587><U+2582><U+2582>"
y[1]: "▃▅▅▇▃▇▂▂"

2. Error: Skimr kable prints as expected, 64-bit (@test-skim_print.R#45) -------
could not find function "skip_if"
1: .handleSimpleError(function (e)
   {
       e$call <- sys.calls()[(frame + 11):(sys.nframe() - 2)]
       register_expectation(e, frame + 11, sys.nframe() - 2)
       signalCondition(e)
   }, "could not find function \"skip_if\"", quote(skip_if(R.Version()$arch == "i386"))) at C:\Users\Maelle\Documents\skimr/tests/testthat/test-skim_print.R:45
2: eval(code, test_env)

3. Failure: make_utf8 produces the correct result  (@test-skim_print.R#183) ----
`input` not identical to `correct`.
1/1 mismatches
x[1]: "<U+2585>"
y[1]: "▅"

4. Failure: skim_v returns expected response for numeric vectors (@test-skim_v.R#19)
`input` not identical to `correct`.
Rows in x but not y: 11. Rows in y but not x: 11.

5. Failure: skim_v handles numeric vectors with NAs and extreme numbers (@test-skim_v.R#73)
`input` not identical to `correct_patho`.
Rows in x but not y: 11. Rows in y but not x: 11.

6. Failure: skim_v returns expected response for ts vectors (@test-skim_v.R#220)
`input` not identical to `correct`.
Rows in x but not y: 13. Rows in y but not x: 13.

7. Failure: inline_linegraph returns expected response for a ts vector (@test-stats.R#6)
`input` not identical to `correct`.
1/1 mismatches
x[1]: "<U+28C0><U+28C0><U+2824><U+2824><U+2812><U+2812><U+2809><U+2809>"
y[1]: "⣀⣀⠤⠤⠒⠒⠉⠉"

8. Failure: inline_linegraph returns expected response for a long ts vector. (@test-stats.R#12)
`input` not identical to `correct`.
1/1 mismatches
x[1]: "<U+28C0><U+28C0><U+28C0><U+2814><U+2824><U+280A><U+2811><U+280A>"
y[1]: "⣀⣀⣀⠔⠤⠊⠑⠊"

9. Failure: inline_linegraph returns expected response for an NA ts vector. (@test-stats.R#18)
`input` not identical to `correct`.
1/1 mismatches
x[1]: "<U+2800>"
y[1]: "⠀"

10. Failure: inline histogram is calculated correctly. (@test-stats.R#46) ------
`input` not identical to `correct`.
1/1 mismatches
x[1]: "<U+2582><U+2587><U+2585><U+2587><U+2586><U+2585><U+2582><U+2582>"
y[1]: "▂▇▅▇▆▅▂▂"

11. Failure: inline histogram is calculated correctly when x is all zeros. (@test-stats.R#53)
`input` not identical to `correct`.
1/1 mismatches
x[1]: "<U+2581><U+2581><U+2581><U+2587><U+2581><U+2581><U+2581><U+2581>"
y[1]: "▁▁▁▇▁▁▁▁"

12. Failure: inline histogram is calculated correctly when x is all 1s. (@test-stats.R#65)
`input` not identical to `correct`.
1/1 mismatches
x[1]: "<U+2581><U+2581><U+2581><U+2587><U+2581><U+2581><U+2581><U+2581>"
y[1]: "▁▁▁▇▁▁▁▁"

DONE ===========================================================================
````

Note: I'll erase this once it's solved in order not to clutter this thread too much.
