+++ tests/testthat/test-skim_print.R
@@ -99,16 +99,16 @@ test_that("Skimr kable prints as expected, 32-bit windows", {
   
   expect_length(input, 18)
   expect_equal(input[15], 
-"|Petal.Length |0       |150      |150 |3.76 |1.77 |1   |1.6 |4.35   |5.1 |6.9  |▇▁▁▂▅▅▃▁ |"
+"|Petal.Length |0       |150      |150 |3.76 |1.77 |1   |1.6 |4.35   |5.1 |6.9   |▇▁▁▂▅▅▃▁ |"
    )
   expect_equal(input[16], 
-"|Petal.Width  |0       |150      |150 |1.2  |0.76 |0.1 |0.3 |1.3    |1.8 |2.5  |▇▁▁▃▃▃▂▂ |"
+"|Petal.Width  |0       |150      |150 |1.2  |0.76 |0.1 |0.3 |1.3    |1.8 |2.5   |▇▁▁▃▃▃▂▂ |"
   )
   expect_equal(input[17], 
-"|Sepal.Length |0       |150      |150 |5.84 |0.83 |4.3 |5.1 |5.8    |6.4 |7.9  |▂▇▅▇▆▅▂▂ |"
+"|Sepal.Length |0       |150      |150 |5.84 |0.83 |4.3 |5.1 |5.8    |6.4 |7.9   |▂▇▅▇▆▅▂▂ |"
    )
   expect_equal(input[18], 
-"|Sepal.Width  |0       |150      |150 |3.06 |0.44 |2   |2.8 |3      |3.3 |4.4  |▁▂▅▇▃▂▁▁ |"
+"|Sepal.Width  |0       |150      |150 |3.06 |0.44 |2   |2.8 |3      |3.3 |4.4   |▁▂▅▇▃▂▁▁ |"
    )
 })
 
