### 2018-01-09 #175 @maelle
The tests now pass on my laptop using the latest `testthat` version, thanks @jimhester for the help!

He suggested to add `testthat (>= 2.0.0)` in DESCRIPTION.
