### 2018-01-09 #175 @elinw
Yes, as of now we have fixed the Windows issue  in master.  I'm about to push one more change and then I'm going to tag 1.0.1 sometime today.  I'll definitely update the testthat minimum first.
