### 2018-01-09 #175 @maelle
And you can confirm that 1.0.1 will be good for review? I even have two reviewers lined up once you confirm that. 😄
