### 2018-01-09 fd44685
Add testthat minimum version.
