### 2018-01-09 #175 @maelle
### Editor checks:

- [x ] **Fit**: The package meets criteria for [fit](policies.md#fit) and [overlap](policies.md#fit)
- [ x] **Automated tests:** Package has a testing suite and is tested via Travis-CI or another CI service.
- [ x] **License:** The package has a CRAN or OSI accepted license
- [ x] **Repository:** The repository link resolves correctly

---

#### Editor comments

Thanks again for your submission, if I understand correctly the version for review will be ready today?

* Add `testthat` minimal version in DESCRIPTION.

* You don't need to fill the Maintainer field in DESCRIPTION, it'll get filled from Authors@R

`goodpractice` output for you and the reviewers to have in mind

```r
It is good practice to

  ✖ write short and simple functions. These functions have high cyclomatic
    complexity:kable.data.frame (60).
  ✖ avoid long code lines, it is bad for readability. Also, many people prefer editor
    windows that are about 80 characters wide. Try make your lines shorter than 80
    characters

    R\skim.R:14:1
    R\skim.R:73:1
    R\stats.R:28:1
    R\stats.R:52:1
    R\stats.R:64:1
    ... and 20 more lines

```

Two typos identified via `devtools::spell_check()`

```r
differnt      skimr-package.Rd:16
faciliates    skim_to_wide.Rd:19
```
---

Reviewers: @jimhester @jenniferthompson
Due date: 2018-01-30
