### 2018-01-09 #175 @maelle
@jimhester and @jenniferthompson thanks for accepting to review this package! Your reviews are due on the 2018-01-30. 😺

You'll find here [the review template](https://github.com/ropensci/onboarding/blob/master/reviewer_template.md) and [our reviewing guide](https://github.com/ropensci/onboarding/blob/master/reviewing_guide.md).
