### 2018-01-09 #175 @elinw
Thanks for the quick checks.  I have done the easy ones, but there is one issue that we can't fix and two I'd like feedback on.

1. The complex function one is a function that is brought in from knitr so we can't do anything about that.  So in our code it is just
`kable.data.frame <- knitr::kable
`

2. In terms of long lines, the ones that remain are in tests of the print function so they are the full printed line.  If we don't do that we get the tibble wrap which defeats the idea of testing the whole line.  On the other hand if we want we could split those strings into two parts and then `paste0()` them which would keep the lines under 80 but still give us a correct version that is the full line.  Does it make sense to do that? Or does that actually make the tests themselves harder to read?

3. We have some things that are only run for windows_os and those are the lines showing up without code coverage in good practices. We've excluded them from testing on Travis since the code will never be hit on Linux or Mac.  Is it possible to write conditional exclusions so that the tests would run when built on Windows?
