### 2018-01-09 #175 @elinw
Cool, just add the same line to appveyor.yml?
`  - Rscript -e 'covr::codecov()'`
