### 2018-01-09 #175 @jimhester
Appveyor needs to use double quotes and you can add it to `on_success:`

```yaml
on_success:
  - Rscript -e "covr::codecov()"
```

And I think you may need to add your codecov token to the appveyor settings. You can get the token at https://codecov.io/gh/ropenscilabs/skimr/settings, and add it as `CODECOV_TOKEN` at https://ci.appveyor.com/project/ropenscilabs/skimr/settings/environment
