### 2018-01-09 c3a14fd
Add covr to Appveyor and delete Maintainer line.
