+++ appveyor.yml
@@ -43,3 +43,6 @@ artifacts:
 
   - path: '\*_*.zip'
     name: Bits
+
+on_success:
+  - Rscript -e "covr::codecov()"
