+++ DESCRIPTION
@@ -42,7 +42,8 @@ Suggests:
     extrafont,
     rmarkdown,
     testthat (>= 2.0.0),
-    withr
+    withr,
+    covr
 License: GPL-3
 Encoding: UTF-8
 LazyData: true
