### 2018-01-10 c478ace
Cleaning up NEWS
Little formatting stuff.
