+++ NEWS.md
@@ -1,16 +1,17 @@
 skimr 1.0.1 (2018-01-xx)
 ========================
 ### NEW FEATURES
-   * Add support for spark plots on Windows
+  * Add support for spark plots on Windows
 
 ### MAJOR CHANGES
-   * spark_line() and spark_bar() are no longer exported
-   * Default statistics for numeric changed from min(x) and max(x) to 
-     quantile(x, probs= c(0)) and quantile(x, probs= c(1))
+  * spark_line() and spark_bar() are no longer exported
+  * Default statistics for numeric changed from min(x) and max(x) to 
+    quantile(x, probs = 0) and quantile(x, probs = 1). These changes
+    lead to more predictable behaviors when a column is all NA values.
 
 #### MINOR IMPROVEMENTS
-   * Add minimimum required version for stringr
-   * Improve documentation related to fonts
+  * Add minimimum required version for stringr
+  * Improve documentation related to fonts
 
 ### BUG FIXES
   * Fix issue where a histogram for data with all NAs threw an error
