### 2018-01-10 dd31047
Refactor skim_to_list. No user-facing changes.
