### 2018-01-10 3e43ef7
Attach magrittr for the example
