+++ R/skim.R
@@ -129,11 +129,11 @@ skim_to_wide <- function(x, ...) {
 #' sl[["numeric"]]
 #' kable(sl$numeric)
 #' 
-#' # Or grouped, using magrittr exposition pipe
+#' # Or grouped, this uses the magrittr exposition pipe
 #' # see ?magrittr::`%$%`
 #' library(magrittr)
 #' iris %>%
-#'   group_by(Species) %>%
+#'   dplyr::group_by(Species) %>%
 #'   skim_to_list() %$%
 #'   kable(numeric)
 #' @export
