### 2018-01-11 2ae4567
Small typos
