### 2018-01-12 dafaf89
Bump version. Update Using Skimr vignette.
