### 2018-01-14 9e0510c
Doc fix
The goals of this process were to increase doc visibility by combining
separate files, provide better cross references and surface aspects
of `skim` that might not be well known.

This commit also fixes the vignettes, which weren't accessible within
R.
