### 2018-01-14 1f58111
Move quiet_impl to wide.R
