+++ DESCRIPTION
@@ -43,8 +43,7 @@ Suggests:
     rmarkdown,
     testthat (>= 2.0.0),
     withr,
-    covr,
-    knitr
+    covr
 License: GPL-3
 Encoding: UTF-8
 LazyData: true
