### 2018-01-14 1627eb8
Removing more Rd formatting.
