+++ R/kable.R
@@ -3,9 +3,9 @@
 #' This is a very simple table generator. It is simple by design. It is not
 #' intended to replace any other R packages for making tables.
 #' 
-#' @seealso \code{\link[knitr]{kable}}
+#' @seealso [`knitr::kable()`]
 #' @inheritParams knitr::kable
-#' @return The original \code{skim_df} object.
+#' @return The original `skim_df` object.
 #' @export
 
 kable <- function (x, format = NULL, digits = getOption("digits"), 
@@ -19,7 +19,7 @@ kable <- function (x, format = NULL, digits = getOption("digits"),
 kable.data.frame <- knitr::kable
 
 
-#' @describeIn kable Produce \code{kable} output of a skimmed data frame
+#' @describeIn kable Produce `kable` output of a skimmed data frame
 #' @export
 
 kable.skim_df <- function(x, format = NULL, digits = getOption("digits"), 
@@ -56,7 +56,7 @@ kable_impl <- function(transformed_df, skim_type, format , digits, row.names,
 }
 
 
-#' @describeIn kable Kable method for a summary_skim_df object
+#' @describeIn kable Kable method for a `summary_skim_df` object
 #' @export
 
 kable.summary_skim_df <- function(x, ...) {
