### 2018-01-18 b294e3a
add na.rm = TRUE to fix #255
