### 2018-01-21 9fc7269
Added tests addressing changes for fixing #255
Added tests to check proper skimming on numeric variables with all zeroes or NAs
