+++ tests/testthat/test-skim_v.R
@@ -237,7 +237,7 @@ test_that("skim_v returns expected response for POSIXct vectors", {
 })
 
 test_that("skim_v returns expected response for list (not AsIs) vectors", {
-  correct <- tibble::tribble (
+  correct <- tibble::tribble(
     ~type,    ~stat,           ~level,  ~value, ~formatted,
     "list",   "missing",        ".all",  1L,     "1",
     "list",   "complete",       ".all",  5L,     "5",
@@ -258,7 +258,7 @@ test_that("skim_v returns expected response for list (not AsIs) vectors", {
 })
 
 test_that("skimr:::skim_v returns expected response for list with all NA's", {
-  correct <- tibble::tribble (
+  correct <- tibble::tribble(
     ~type,    ~stat,           ~level,  ~value, ~formatted,
     "list",   "missing",       ".all",   3L,     "3",
     "list",   "complete",      ".all",   0L,     "0",
@@ -329,5 +329,24 @@ test_that("numeric skim is calculated correctly when x is all NAs.", {
     "numeric",      "p75",   ".all", NA,                   "NA",         "x",
     "numeric",      "p100",  ".all", NA,                    "NA",         "x",
     "numeric",     "hist",   ".all", NA,                   " ",         "x" )
+  expect_identical(input[1:6], correct[1:6])
+})
+
+test_that("numeric skim is calculated correctly when x is all zeores or NAs.", {
+  x <- as.numeric(c(NA, NA, NA, 0))
+  input <- skim(x)
+  correct <- tibble::tribble(
+    ~type,          ~stat, ~level,   ~value,              ~ formatted, ~variable,
+    "numeric",  "missing",   ".all", 3,                   "3",         "x",
+    "numeric", "complete",   ".all", 1,                   "1",         "x",
+    "numeric",        "n",   ".all", 4,                   "4",         "x",
+    "numeric",     "mean",   ".all", 0,                   "0",         "x",
+    "numeric",       "sd",   ".all", NA,                  "NA",         "x", 
+    "numeric",      "p0",   ".all",  0,                   "0",         "x",
+    "numeric",      "p25",   ".all", 0,                   "0",         "x",
+    "numeric",   "median",   ".all", 0,                   "0",         "x",
+    "numeric",      "p75",   ".all", 0,                   "0",         "x",
+    "numeric",      "p100",  ".all", 0,                   "0",         "x",
+    "numeric",     "hist",   ".all", NA,  "▁▁▁▇▁▁▁▁",         "x" )
   expect_identical(input[1:6], correct[1:6])
 })
