### 2018-01-21 557df01
Start work on updating tests
