### 2018-01-22 80b1392
Make vector output match data frame output more closely.
