+++ R/skim.R
@@ -64,7 +64,10 @@ skim.default <-function(.data, ...){
   }
   skimmed <- skim_v(.data)
   skimmed$variable <- deparse(substitute(.data))
-  structure(skimmed, class = c("skim_vector", class(skimmed)) )
+  skimmed <- dplyr::select(skimmed, variable, type, stat, level, 
+                           value, formatted)
+  structure(skimmed, class = c("skim_vector", "skim_df", class(skimmed)),
+            df_name = skimmed$variable[1])
 }
 
 
