### 2018-01-28 2247fac
Support simultaneously assignment and removal of skimmers.
