+++ R/functions.R
@@ -33,6 +33,17 @@ NULL
 #' skim_with(numeric = list(hist = NULL))
 #' skim(faithful)
 #' 
+#' # This works with multiple skimmers. Just match names to overwrite
+#' skim_with(numeric = list(iqr = IQR, p25 = NULL, p75 = NULL))
+#' skim(faithful)
+#' 
+#' # Alternatively, set `append = FALSE` to replace the skimmers of a type.
+#' skim_with(numeric = list(mean = mean, sd = sd), append = FALSE)
+#' 
+#' # Skimmers are unary functions. Partially apply arguments during assigment.
+#' # For example, you might want to remove NA values.
+#' skim_with(numeric = list(iqr = purrr::partial(IQR, na.rm = TRUE)))
+#' 
 #' # Go back to defaults
 #' skim_with_defaults()
 #' skim(faithful)
