### 2018-01-28 f8891f3
Support formula-style lambdas from rlang.
