### 2018-01-28 26c7898
Delete duplicate entry in news
