### 2018-01-28 979bb79
Merge changes from skim_with branch.
