### 2018-01-29 #175 @jenniferthompson
Also, if I've caught some typos in the vignettes, would y'all like a PR or no?
