### 2018-01-29 #175 @jimhester
---

All review comments are based off of version 1.0.1. Overall I feel like the
package would be a welcome addition to rOpenSci after the review process is complete.

  - Get rid of `## Dev mode: on` at the top.
  - devel installation instructions better written as
    `devtools::install_github("ropenscilabs/skimr@develop")` (not sure it is
    necessary to mention the develop branch at all here)
  - A markdown file should not have a yaml header and this file should really
    be built from a README.Rmd included in the repository.
    `usethis::use_readme_rmd()` is a good way to set this up.
  - Should have links to `kable()` and `pander()`, and possibly include
    examples of what 'enhanced' options are available and what they look like.

### skim

Displaying the results of `skim()` as seemingly a  table, but having the
underlying data as a table in a different format is quite surprising. I pretty
strongly think that you should have two functions, one which returns the
summary information in a wide format, the other in the long format. e.g. From
the print output I would expect to be able to run the following to retrieve
variables in the dataset with standard deviations greater than 1

```r
skim(iris, Sepal.Length, Petal.Length) %>%
  filter(sd > 1) %>%
  pull(variable)
```

but this instead throws a pretty confusing error

```
Error in filter_impl(.data, quo) :
  Evaluation error: comparison (6) is possible only for atomic and list types.
```

Because the data is actually in long format.

```r
skim(iris, Sepal.Length, Petal.Length) %>%
  filter(stat == "sd", value > 1) %>%
  pull(variable)
```

Gives you the expected output.

As the print method hides this dichotomy I think this will cause users
confusion which would be best remedied by having two different functions. A
simple solution would be to return the `skim_to_wide()` representation from
`skim()`. Or maybe instead have a series of functions

- skim_wide()
- skim_long()
- skim_list()

Regardless of the exact API I think you should think hard about making it more
obvious the results of `skim()` are not simple rectangular tables.

I realize this is somewhat challenging because you have separate output for
different column types. Another possible solution which would require minimal
API changes is to modify the print method to make the heterogeneous parts more
clear.

### skim_with

Designing `skim_with()` as a impure function which modifies global state seems
an odd choice when it could instead be written as a functional which returned
`skim()` functions with alternate defaults. Beside the impurity concerns it
also makes it more difficult to use different types of skim functions
interchangeably. If you really want the impure interface you could do so by
defining `skim_with_defaults()` with the same interface as `skim_with()` and an
implementation that simply calls the pure `skim_with()` function and assigns
the result as the new default.

### Vignettes

Having the 'Using Fonts' vignette be a rmarkdown template is kind of strange,
also I think it might be worth listing a few fonts which _do_ support all the
features needed for the sparklines / historgrams would be useful.

### Miscellanea

[get_funs()](https://github.com/ropenscilabs/skimr/blob/ee10b781ad7cc935437a22dd51f32fc38cfba70d/R/functions.R#L106-L109) has a very weird implementation, particularly because it's behavior seems equivalent to double square bracket. e.g.
`options$functions[["numeric"]]` gives equivalent output to `get_func("numeric")`, likewise for `get_func("xyz")` (which is `NULL`).

The code [inherited from
pillar](https://github.com/ropenscilabs/skimr/commit/cc75bb4f083c0831a9f7fab0a451427f15367e5e)
should have a comment nearby indicating such, so that you can more easily track
upstream changes if the pillar implementation changes in the future.

I really think you should reconsider changing `min` and `max` to `p0` and
`p100`. `min()` and `max()` are self explanitory, but I would need to read the
quantile documentation to be sure that `p0` is guaranteed to be the minimum
value in the distribution not subject to rounding error. I would actually
prefer dealing with `-Inf` and `Inf` in the all NA case to using
`quantile(probs = 0)`.
