### 2018-01-30 #175 @jenniferthompson
## skimr Package Review
### Jennifer Thompson

*This review is of `skimr` version 1.0.2, downloaded from Github 2018-01-29. Other relevant info: R version is 3.4.3; devtools version is 1.13.4.*

- [x] As the reviewer I confirm that there are no conflicts of interest for me to review this work.

- [x] **Examples** for all exported functions in R Help that run successfully locally *Yes, but see below*

---

skimr is a very, very useful package that is generally well thought in terms of both simplicity and extensibility. I particularly appreciate a thorough vignette for the main functions, intuitive function names, and the attention to where many users are likely to run into trouble, along with the proactive explanation of potential solutions. skimr does a specific set of tasks and does them simply and well. The source code is easy to digest in both style and function, and the comments are abundant, clear and helpful.

I do agree with Jim that it's confusing to have the printed output be a completely different format from what is output to the console; I'm most excited about (and expect) the wide format printed, vs the long data.frame returned. Perhaps have the default `skim()` return a named list, with an element for each data type containing the "wide" format as a data.frame, and have a `skim_long` that returns the current long data frame.

Some specific suggestions:

- README:
    - Clarify first bullet - "added missing, complete, n, sd" to what?
    - Suggested edit for brevity to statement of need: *"skimr provides a frictionless approach to summary statistics which conforms to the principle of 'least surprise' `[link to description of this principle]`, displaying summary statistics the user can skim quickly to understand their data. It handles different data types and returns a `skimr` object which can be included in a pipeline or displayed nicely for the human reader."*
    - Remove YAML output at the top.
- Really helpful vignette for `skim()` - great work showing common use cases and some really helpful functionality. Made it easy to add my own summary statistics. I did find a few typos and can include these in a PR if desired.
- Examples in help files:
    - I'm not sure `skim_format(.levels = list(nchar = 4))` is working. I changed the value of this argument several times, and `skim(iris)` looked the same every time, both printed and as a data.frame. Could be user error but I couldn't figure out how to make it affect the output.
    - `?skim:` `group_by` example doesn't actually group by anything (works when I group by Species)
    - I'm not clear on the benefit of having the fonts vignette as an RMarkdown template. Also, it's not showing up as an option for me in RStudio (I only have HTML vignette and Github document, even with skimr loaded).
    - Would be helpful to show an example of `skim_tee()` - I'm not totally clear on when it's beneficial.
- The inline histograms are really helpful for truly continuous variables. Is it possible to make them smarter/more flexible? When I skim a numeric discrete variable with 5 levels (`skim(sample(c(1:5, NA), size = 20, replace = TRUE))`), I get a histogram with what appears to be 8 bars.
- Community guidelines: Provide links to contributing.md, Google style guide and CONDUCT.md.
- The error message when a variable has infinite values (produced via `dplyr::cut`, I think) could be clearer - could be a good warning message to include up front.
- I'd prefer `p50` to `median`, particularly if you stick with `p0` and `p100`. (In terms of workflow, I do prefer NAs to -Inf/Inf, though :) )

Overall I'm excited to use this package in my own work, and really appreciate the hard and thoughtful work of all the authors.
