### 2018-01-31 #175 @maelle
@elinw I had forgotten to ask you to add the rOpenSci review badge to the README! 🙈
