### 2018-01-31 2554c2b
Update readme file based on onboarding feedback.
