+++ inst/other_docs/README.Rmd
@@ -1,146 +1,0 @@
----
-output: 
-  html_document: 
-    keep_md: yes
----
-# skimr
-```{r set-options, echo=FALSE}
-devtools::dev_mode()
-library(skimr)
-skim_with_defaults()
-options(tibble.width = Inf)
-options(width = 120)
-```
-
-[![Build Status](https://travis-ci.org/ropenscilabs/skimr.svg?branch=master)](https://travis-ci.org/ropenscilabs/skimr)
-[![codecov](https://codecov.io/gh/ropenscilabs/skimr/branch/master/graph/badge.svg)](https://codecov.io/gh/ropenscilabs/skimr)
-
-The goal of skimr is to provide a frictionless approach to dealing with summary statistics iteratively and interactively as part of a pipeline, and that conforms to the principle of least surprise. 
-
-`Skimr` provides summary statistics that you can skim quickly to understand and your data and see what may be missing. It handles different data types (numerics, factors, etc), and returns a skimr object that can be piped or displayed nicely for the human reader. 
-
-## Installation
-
-
-``` r
-# install.packages("devtools")
-devtools::install_github("ropenscilabs/skimr")
-```
-
-To install the version with the most recent changes that have not yet been incorporated in 
-the master branch (and may not be):
-
-``` r
-devtools::install_github("ropenscilabs/skimr", ref = "develop")
-```
-
-## Skim statistics in the console
-
-- added missing, complete, n, sd
-- reports numeric/int/double separately from factor/chr
-- handles dates, logicals
-- supports spark-bar and spark-line based on 
-[Hadley Wickham's pillar package](https://github.com/hadley/pillar).
-
-**Nicely separates variables by class:**  
-
-```{r}
-skim(chickwts)
-```
-
-**Presentation is in a compact horizontal format:**  
-```{r}
-skim(iris)
-```
-
-Individual columns of a data frame can be selected using tidyverse style selectors.
-
-```{r}
-skim(iris, Sepal.Length, Petal.Length)
-```
-
-## Handles grouped data
-
-Skim() can handle data that has been grouped using `dplyr::group_by`.
-
-```{r}
-iris %>% dplyr::group_by(Species) %>% skim()
-```
-
-## Options for kable and pander
-
-Enhanced print options are available by piping to kable() or pander(). 
-
-## skim_df object (long format)
-
-By default `skim` prints beautifully in the console, but it also produces a long, tidy-format skim_df object that can be computed on. 
-
-```{r}
-a <-  skim(chickwts)
-dim(a)
-```
-
-```{r}
-print.data.frame(skim(chickwts))
-```
-
-## Compute on the full skim_df object
-```{r}
-skim(mtcars) %>% dplyr::filter(stat=="hist")
-```
-
-## Works with strings, lists and other column classes.
-
-```{r}
-skim(dplyr::starwars)
-```
-
-Users can add new classes.
-
-## Specify your own statistics
-
-```{r}
-funs <- list(iqr = IQR,
-    quantile = purrr::partial(quantile, probs = .99))
-  skim_with(numeric = funs, append = FALSE)
-  skim(iris, Sepal.Length)
-# Restore defaults
-  skim_with_defaults()
-```
-
-## Limitations of current version
-
-We are aware that there are issues with rendering the inline histograms and line charts in various contexts, some of which are 
-described below. 
-
-### Windows support for spark histograms
-
-Windows cannot print the spark-histogram characters when printing a data-frame. For example, 
-`"▂▅▇"` is printed as `"<U+2582><U+2585><U+2587>"`. This longstanding problem [originates in 
-the low-level code](http://r.789695.n4.nabble.com/Unicode-display-problem-with-data-frames-under-Windows-td4707639.html) 
[…]
