### 2018-01-31 #175 @elinw
Oops   @jenniferthompson also mentioned the template (and had the same issue as someone in the issue tracker) so I'm trying to hunt that down. It is usually something goofy when they don't show up.

I think that with the histograms we need to come up with a more structured and probably less elegant approach when there are 2-9 separate values.  We were more focused on the too many case than the too few.

I'm not sure what is going on with the help index, on my rstudioserver install from CRAN I have
![screen shot 2018-01-31 at 12 17 58 pm](https://user-images.githubusercontent.com/754813/35640454-329fd82a-068b-11e8-9731-ef7add25f56a.png)

![screen shot 2018-01-31 at 12 18 08 pm](https://user-images.githubusercontent.com/754813/35640441-276a4a94-068b-11e8-886c-c1c8f15f5d97.png)

But definitely don't have that when I build locally.
