+++ inst/rmarkdown/templates/fonts-in-skimr/template.yaml
@@ -1,3 +1,3 @@
-name: Fonts in skimr
+name: Fonts_in_skimr
 description: >
  Using fonts in skimr
