### 2018-02-03 #175 @elinw
Some of the items listed are now in master (basic fixes to the readme, make the template work, add a bit to the fonts vignette, fixed the wrong code in the format documentation).  @jenniferthompson About Inf values, you're talking about when the data being skimmed has Inf as opposed to when the skim produces it?  That's definitely something we need to look at and handle. Probably overall edge case testing needs to b examined.  I have a refactoring of tests project I'm working on so I'll include the cases I can think of.
