### 2018-02-05 #175 @elinw
Well we count NAs so it would potentially make sense to count Inf, -Inf and
maybe NaN also.

On Sat, Feb 3, 2018 at 5:20 PM, Jennifer Thompson <notifications@github.com>
wrote:

> About Inf values, you're talking about when the data being skimmed has Inf
>
> @elinw <https://github.com/elinw> Yes, exactly! I tried it on one of my
> datasets; my variable really shouldn't have had Inf as a value, so it was
> helpful in that it made me aware of it once I figured out what variable was
> causing the error. But I was thinking that maybe even more helpful would be
> either a check up front for Inf values with a more specific error message,
> or a workaround in the code to avoid the error but an informative warning
> message.
>
