### 2018-02-12 8cfc863
Edit README summary statement, plus typo/flow edits
