+++ README.Rmd
@@ -15,27 +15,17 @@ options(width = 120)
 [![codecov](https://codecov.io/gh/ropenscilabs/skimr/branch/master/graph/badge.svg)](https://codecov.io/gh/ropenscilabs/skimr)
 [![](https://badges.ropensci.org/175_status.svg)](https://github.com/ropensci/onboarding/issues/175)
 
-The goal of skimr is to provide a frictionless approach to dealing with 
-summary statistics iteratively and interactively as part of a pipeline, 
-and that conforms to the principle of least surprise. 
-
-`Skimr` provides summary statistics that you can skim quickly to understand 
-your data and see what may be missing. It handles different data types 
-(numerics, factors, etc), and returns a skimr object that can be piped or 
-displayed nicely for the human reader. 
+`skimr` provides a frictionless approach to summary statistics which conforms to the [principle of least surprise](https://en.wikipedia.org/wiki/Principle_of_least_astonishment), displaying summary statistics the user can skim quickly to understand their data. It handles different data types and returns a `skim_df` object which can be included in a pipeline or displayed nicely for the human reader.
 
 ## Installation
 
-The current released version of Skimr can be install from CRAN. 
-If you wish to install the current build of the next release you can do so 
-using the following:
+The current released version of `skimr` can be installed from CRAN. If you wish to install the current build of the next release you can do so using the following:
 
 ``` r
 # install.packages("devtools")
 devtools::install_github("ropenscilabs/skimr")
 ```
-The APIs for this branch should be considered reasonably stable 
-but still subject to change if an issue is discovered.
+The APIs for this branch should be considered reasonably stable but still subject to change if an issue is discovered.
 
 To install the version with the most recent changes that have not yet been 
 incorporated in the master branch (and may not be):
@@ -47,9 +37,11 @@ Do not rely on APIs from the develop branch.
 
 ## Skim statistics in the console
 
-- Provides a larger set of statistics than `summary()` including missing, 
+`skimr`:
+
+- Provides a larger set of statistics than `summary()`, including missing, 
 complete, n, sd
-- reports numeric/int/double separately from factor/chr
+- reports numeric/int/dbl types separately from factor/chr
 - handles dates, logicals
 - supports spark-bar and spark-line based on 
 [Hadley Wickham's pillar package](https://github.com/hadley/pillar).
@@ -71,21 +63,21 @@ skim(iris)
 skim(iris, Sepal.Length, Petal.Length)
 ```
 
-** Handles grouped data: **
+**Handles grouped data: **
 
-Skim() can handle data that has been grouped using `dplyr::group_by`.
+`skim()` can handle data that has been grouped using `dplyr::group_by`.
 
 ```{r}
 iris %>% dplyr::group_by(Species) %>% skim()
 ```
 
-** Options for kable and pander: **
+**Options for kable and pander: **
 
-Enhanced print options are available by piping to kable() or pander(). 
+Enhanced print options are available by piping to `kable()` or `pander()`. 
 
-## skim_df object (long format)
+## `skim_df` object (long format)
 
-By default `skim` prints beautifully in the console, but it also produces a long, tidy-format skim_df object that can be computed on. 
+By default `skim()` prints beautifully in the console, but it also produces a long, tidy-format `skim_df` object that can be computed on. 
 
 ```{r}
 a <-  skim(chickwts)
@@ -96,12 +88,12 @@ dim(a)
 print.data.frame(skim(chickwts))
 ```
 
-## Compute on the full skim_df object
+## Compute on the full `skim_df` object
 ```{r}
 skim(mtcars) %>% dplyr::filter(stat=="hist")
 ```
 
-## Works with strings, lists and other column classes.
[…]
