### 2018-02-12 ceafeab
Edits for typos, clarity/flow
