+++ vignettes/Using_skimr.Rmd
@@ -9,12 +9,9 @@ vignette: >
   %\VignetteEncoding{UTF-8}
 ---
 
-The skimr package is designed to provide summary statistics about variables. In
-base R the most similar functions are `summary()` for vectors and data frames
-and `fivenum()` for numeric  vectors. Skimr is opinionated in its defaults but
-easy to modify.
+`skimr` is designed to provide summary statistics about variables. It is opinionated in its defaults, but easy to modify.
 
-For comparison purposes here are examples of the similar functions.
+In base R, the most similar functions are `summary()` for vectors and data frames and `fivenum()` for numeric  vectors:
 
 ```{r}
 summary(iris)
@@ -29,62 +26,44 @@ fivenum(iris$Sepal.Length)
 summary(iris$Species)
 ```
 
-# The skim function
+# The `skim` function
 
-The core function of skimr is skim().  Skim is a S3 generic function; the skim
-package  includes support for data frames, grouped data frames and vectors 
-(for a number of standard types of data). Like summary
-for data frames, skim presents results for all of the columns and the statistics
-depend on the class of the variable.
+The core function of skimr is `skim()`. `skim()` is a S3 generic function, with methods for data frames, grouped data frames and vectors. Like `summary()`, `skim()`'s method for data frames presents results for every column; the statistics it provides depend on the class of the variable.
 
 ## Skimming data frames
 
-The main focus of skimr is on data frames. This is by design since it relies 
-extensively on the tidyverse vocabulary and is pipeable and those methods 
-are designed for data frames. 
+By design, the main focus of `skimr` is on data frames; it is intended to fit well withiin a data [pipeline](http://r4ds.had.co.nz/pipes.html) and relies extensively on [tidyverse](https://www.tidyverse.org/) vocabulary, which focuses on data frames.
 
-Unlike summary.data.frame(), the printed results of skim (those displayed in
-the console or in a knitted markdown file) are shown horizontally with one row
-per variable and separated into separate tibbles for each class of variables.
-The actual results are  stored in a `skim_df` object that is also a tibble 
-In `summary.data.frame()` the statistics are stored in a `table` with
-one column for each variable and the standard table printing is used to display
-the results. 
+Results of `skim()` are *printed* horizontally, with one section per variable type and one row per variable. Results are *returned* from `skim()` as a long [tibble](http://tibble.tidyverse.org/) of class `skim_df`, with one row per variable + summary statistic.
 
 ```{r}
 library(skimr)
 skim(iris)
 ```
 
-This distinction is important because the skim_df object is easy to use for
-additional manipulation if desired and is pipeable. For example all of the
-results for a particular statistic or for one  variable could be selected or
-an alternative printing method sould be developed.  
+This is in contrast to `summary.data.frame()`, which stores statistics in a `table`. The distinction is important, because the `skim_df` object is pipeable and easy to use for additional manipulation: for example, the user could select all of the variable means, or all summary statistics for a specific variable.
+
+```{r}
+skim(iris) %>%
+  dplyr::filter(stat == "mean")
+
+```
 
 The `skim_df` object always contains 6 columns: 
 
-- `variable` the name of the original variable
-- `type` the class of the variable
-- `stat` the statistic calculated with the name becoming the column name when
-  the object is printed.
-- `level` is used when summary functions returns multiple values when skimming.
-  This happens as for counts of levels for factor variables or when setting
-  multiple values to the `probs` argument of the `quantiles` function.
-- `value` is the actual calculated value of the statistic and should be used
-  for further calculations. This is always numeric.
-- `formatted` is a formatted version of value that attempts to use
[…]
