### 2018-02-12 ropenscilabs/skimr#279 @jenniferthompson
Clarification/typos in docs
I found a few typos in the README and "using skimr" vignette, and will open a PR as discussed in the [onboarding discussion](https://github.com/ropensci/onboarding/issues/175) with some fixes and edits for clarity. Happy for anything to be used or not as y'all like; let me know if you'd like anything in the process done differently!
