### 2018-02-12 d8fdd01
Fix missing grouping in documentation
Thanks for pointing this out @jenniferthompson!
