+++ R/skim.R
@@ -51,7 +51,7 @@ globalVariables(".")
 #' skim(iris, starts_with("Sepal"))
 #' 
 #' # Skim also works groupwise
-#' dplyr::group_by(iris) %>% skim()
+#' dplyr::group_by(iris, Species) %>% skim()
 #' 
 #' # Skim pipelines; now we work with the tall format
 #' skim(iris) %>% as.data.frame()
