### 2018-02-12 #175 @michaelquinn32
Thanks for the reviews everyone! I've tried to reduce it to a set of issues that we can start tracking. Please let me know if I missed anything:

1. Fix readme. The bullet points are clear.
    * Get rid of ## Dev mode: on at the top.
    * devel installation instructions better written as
    devtools::install_github("ropenscilabs/skimr@develop") (not sure it is
    necessary to mention the develop branch at all here)
    * A markdown file should not have a yaml header and this file should really
    be built from a README.Rmd included in the repository.
    usethis::use_readme_rmd() is a good way to set this up.
    * Should have links to kable() and pander(), and possibly include
    examples of what 'enhanced' options are available and what they look like.
  * Clarify first bullet - "added missing, complete, n, sd" to what?
   * Suggested edit for brevity to statement of need: "skimr provides a frictionless approach to summary statistics which conforms to the principle of 'least surprise' [link to description of this principle], displaying summary statistics the user can skim quickly to understand their data. It handles different data types and returns a skimr object which can be included in a pipeline or displayed nicely for the human reader."

2. Coalesce skim api around skim_long (aliased skim), skim_wide and skim list. The latter two should be support downstream computation, and not just the rendered values. This is going to be a bit tricking for stats with multiple levels, since we normally collapse them in the rendered output.

skim_long is considered the default because it is the most terse representation of relatively heterogeneous data. For any df with more than one data type, skim_wide will be very sparse. Jim's example is going to give one or more rows with lots of NA values where statistics weren't computed.

Skimming to a list is nice for some work but it doesn't play nice with magrittr/dplyr pipes. The other alternative, using nested data frames as a list within a single skimmed data frame, isn't a great idea until dplyr better supports accessing nested data.

```
x <- data.frame(id = 1:3, listcol = I(replicate(3, data.frame(a = 1, b =2), simplify = FALSE)))
dplyr::select(x, listcol$a)
# Throws an error
```

3. Support something in the print that talks about the returned data object being long, wide or a list.
4. (short-term) Rewrite skim_with_defaults() to use the same interface as skim_with(). Same for formatting functions.
5. (long-term) Rewrite skim_with to support a functional  implementation that does not depend on a global state.
6. Better identify code from pillar. Link to it.
7. Update get_funs documentation
8. Fix group_by in skim docs (done). Thanks for pointing that out!
9. Add a skim_tee example
10. Support inline histograms for vectors with less than the default number of characters.
11. Link to contributing.md, tidyverse style guide (replacing Google Style guide) and CONDUCT.md.
12. Rename median to p50, compute using quantile for consistency.

-- Push back
1. get_funs is vectorized, to work with S3 inheritance, which is just a vector of different classes. TBH this is all a little hacky, since we're basically reimplementing S3, but I would love a better idea that allows for dynamic registration of types.

The case that you gave above didn't work for us when we ran into ordered factors.

```
y <- ordered(letters)
class(y)
#> [1] "ordered" "factor"
z <- list(factor = "blue")
z[[class(y)]]
#> Error in z[[class(y)]] : no such index at level 1
```

2. The default value of min for summary is also P0. It uses quantile to produce a five-number summary. I think it's appropriate to go this route as long as the returned statistics are consistently named, which they are. https://github.com/wch/r-source/blob/af7f52f70101960861e5d995d3a4bec010bc89e6/src/library/base/R/summary.R#L42
3. Vignettes, +1 to what Elin already said.
4. For formatting the acceptable values for the .levels options are `max_char = 3` and  `max_levels = 4`. All apologies if the documentation suggested otherwise.
