### 2018-02-12 #175 @elinw
We've fixed a bunch of these issues,.  @michaelquinn32 about the levels issue, the way it can work is really what you get when you go wide (spread) from the long, because the levels are already separate rows in the long form.  I'm just wondering if it makes sense to think about why people want to do calculations, if indeed that is what they want to do.  I think _some_ people want the formatted because they are working on display, while other people want the numeric because they want to do things like filter based on sd or having non missing cases etc.  Before we start implementing something we should think about what the ultimate goal is.  That's why the question of doing selective skimming for a particular type is important too, because what does it mean when you have a mix of types?  By the way the same whole list of concerns applies to many summary.x() methods so its good to try to think this through especially about what is not satisfying in summary() and its associated print methods.  Just as an example, the `print()`  for summary(lm(y~x))  really does not reflect what is in the summary object.  We could even consider a complex object like that.
