### 2018-02-12 #175 @maelle
@elinw @michaelquinn32 thanks both for your work! 😸

Please create related issues in `skimr` repo and report back here your answers to reviewers and the specific questions you may have for them once you're ready. 😉
