+++ README.Rmd
@@ -15,19 +15,26 @@ options(width = 120)
 [![codecov](https://codecov.io/gh/ropenscilabs/skimr/branch/master/graph/badge.svg)](https://codecov.io/gh/ropenscilabs/skimr)
 [![](https://badges.ropensci.org/175_status.svg)](https://github.com/ropensci/onboarding/issues/175)
 
-`skimr` provides a frictionless approach to summary statistics which conforms to the [principle of least surprise](https://en.wikipedia.org/wiki/Principle_of_least_astonishment), displaying summary statistics the user can skim quickly to understand their data. It handles different data types and returns a `skim_df` object which can be included in a pipeline or displayed nicely for the human reader.
+`skimr` provides a frictionless approach to summary statistics which conforms
+to the [principle of least surprise](https://en.wikipedia.org/wiki/Principle_of_least_astonishment),
+displaying summary statistics the user can skim quickly to understand their
+data. It handles different data types and returns a `skim_df` object which can
+be included in a pipeline or displayed nicely for the human reader.
 
 ## Installation
 
-The current released version of `skimr` can be installed from CRAN. If you wish to install the current build of the next release you can do so using the following:
+The current released version of `skimr` can be installed from CRAN. If you wish
+to install the current build of the next release you can do so using the
+following:
 
 ``` r
 # install.packages("devtools")
 devtools::install_github("ropenscilabs/skimr")
 ```
-The APIs for this branch should be considered reasonably stable but still subject to change if an issue is discovered.
+The APIs for this branch should be considered reasonably stable but still
+subject to change if an issue is discovered.
 
-To install the version with the most recent changes that have not yet been 
+To install the version with the most recent changes that have not yet been
 incorporated in the master branch (and may not be):
 
 ``` r
@@ -39,11 +46,11 @@ Do not rely on APIs from the develop branch.
 
 `skimr`:
 
-- Provides a larger set of statistics than `summary()`, including missing, 
+- Provides a larger set of statistics than `summary()`, including missing,
 complete, n, sd
 - reports numeric/int/dbl types separately from factor/chr
 - handles dates, logicals
-- supports spark-bar and spark-line based on 
+- supports spark-bar and spark-line based on
 [Hadley Wickham's pillar package](https://github.com/hadley/pillar).
 
 **Nicely separates variables by class:**  
@@ -57,7 +64,8 @@ skim(chickwts)
 skim(iris)
 ```
 
-**Individual columns of a data frame can be selected using tidyverse-style selectors:**
+**Individual columns of a data frame can be selected using tidyverse-style
+selectors:**
 
 ```{r}
 skim(iris, Sepal.Length, Petal.Length)
@@ -77,7 +85,8 @@ Enhanced print options are available by piping to `kable()` or `pander()`.
 
 ## `skim_df` object (long format)
 
-By default `skim()` prints beautifully in the console, but it also produces a long, tidy-format `skim_df` object that can be computed on. 
+By default `skim()` prints beautifully in the console, but it also produces a
+long, tidy-format `skim_df` object that can be computed on. 
 
 ```{r}
 a <-  skim(chickwts)
@@ -119,13 +128,16 @@ skim_with_defaults()
 
 ## Limitations of current version
 
-We are aware that there are issues with rendering the inline histograms and line charts in various contexts, some of which are described below. 
+We are aware that there are issues with rendering the inline histograms and
+line charts in various contexts, some of which are described below. 
 
 ### Support for spark histograms
 
-There are known issues with printing the spark-histogram characters when printing a data frame. For example, `"▂▅▇"` is printed as `"<U+2582><U+2585><U+2587>"`. This longstanding problem [originates in 
-the low-level code](http://r.789695.n4.nabble.com/Unicode-display-problem-with-data-frames-under-Windows-td4707639.html) 
[…]
