+++ vignettes/Using_skimr.Rmd
@@ -9,9 +9,11 @@ vignette: >
   %\VignetteEncoding{UTF-8}
 ---
 
-`skimr` is designed to provide summary statistics about variables. It is opinionated in its defaults, but easy to modify.
+`skimr` is designed to provide summary statistics about variables. It is
+opinionated in its defaults, but easy to modify.
 
-In base R, the most similar functions are `summary()` for vectors and data frames and `fivenum()` for numeric  vectors:
+In base R, the most similar functions are `summary()` for vectors and data
+frames and `fivenum()` for numeric  vectors:
 
 ```{r}
 summary(iris)
@@ -28,20 +30,32 @@ summary(iris$Species)
 
 # The `skim` function
 
-The core function of skimr is `skim()`. `skim()` is a S3 generic function, with methods for data frames, grouped data frames and vectors. Like `summary()`, `skim()`'s method for data frames presents results for every column; the statistics it provides depend on the class of the variable.
+The core function of skimr is `skim()`. `skim()` is a S3 generic function, with
+methods for data frames, grouped data frames and vectors. Like `summary()`,
+`skim()`'s method for data frames presents results for every column; the
+statistics it provides depend on the class of the variable.
 
 ## Skimming data frames
 
-By design, the main focus of `skimr` is on data frames; it is intended to fit well withiin a data [pipeline](http://r4ds.had.co.nz/pipes.html) and relies extensively on [tidyverse](https://www.tidyverse.org/) vocabulary, which focuses on data frames.
+By design, the main focus of `skimr` is on data frames; it is intended to fit
+well withiin a data [pipeline](http://r4ds.had.co.nz/pipes.html) and relies
+extensively on [tidyverse](https://www.tidyverse.org/) vocabulary, which
+focuses on data frames.
 
-Results of `skim()` are *printed* horizontally, with one section per variable type and one row per variable. Results are *returned* from `skim()` as a long [tibble](http://tibble.tidyverse.org/) of class `skim_df`, with one row per variable + summary statistic.
+Results of `skim()` are *printed* horizontally, with one section per variable
+type and one row per variable. Results are *returned* from `skim()` as a long
+[tibble](http://tibble.tidyverse.org/) of class `skim_df`, with one row per
+variable + summary statistic.
 
 ```{r}
 library(skimr)
 skim(iris)
 ```
 
-This is in contrast to `summary.data.frame()`, which stores statistics in a `table`. The distinction is important, because the `skim_df` object is pipeable and easy to use for additional manipulation: for example, the user could select all of the variable means, or all summary statistics for a specific variable.
+This is in contrast to `summary.data.frame()`, which stores statistics in a
+`table`. The distinction is important, because the `skim_df` object is pipeable
+and easy to use for additional manipulation: for example, the user could select
+all of the variable means, or all summary statistics for a specific variable.
 
 ```{r}
 skim(iris) %>%
@@ -53,17 +67,24 @@ The `skim_df` object always contains 6 columns:
 
 - `variable`: name of the original variable
 - `type`: class of the variable
-- `stat`: name of the summary statistic *(becomes the column name when the object is printed)*
-- `level`: used when summary functions returns multiple values when skimming; for example, counts of levels for factor variables, or when setting multiple values to the `probs` argument of the `quantiles` function
-- `value`: actual calculated value of the statistic; always numeric and should be used for further calculations
-- `formatted`: **formatted** character version of `value`; attempts to use a reasonable number of digits (decimal aligned) and puts values like dates into human readable formats
+- `stat`: name of the summary statistic *(becomes the column name when the
+object is printed)*
+- `level`: used when summary functions returns multiple values when skimming;
+for example, counts of levels for factor variables, or when setting multiple
+values to the `probs` argument of the `quantiles` function
[…]
