### 2018-02-13 d8a6de6
Allow histogram with infinite values
- Replace `seq(..., length = )` with `seq(..., length.out = )` to exactly match the argument.
- Fix issues with histograms of infinite or NA values
- Add tests for those histograms
- Remove a bunch of trailing whitespaces because of Atom
