+++ tests/testthat/test-skim.R
@@ -2,26 +2,26 @@ context("Skim a data.frame")
 
 test_that("Skimming a data frame works as expected", {
   input <- skim(chickwts)
-  
+
   # dimensions
   expect_length(input, 6)
   expect_equal(nrow(input), 23)
-  
+
   # classes
   expect_is(input, "skim_df")
   expect_is(input, "tbl_df")
   expect_is(input, "tbl")
   expect_is(input, "data.frame")
-  expect_identical(input$variable, c(rep(c("weight"), each = 11), 
+  expect_identical(input$variable, c(rep(c("weight"), each = 11),
                                 rep("feed", each = 12)))
-  expect_identical(input$type, c(rep("numeric", each = 11), 
+  expect_identical(input$type, c(rep("numeric", each = 11),
                                  rep("factor", each = 12)))
   expect_identical(head(input$stat),
                    c("missing", "complete", "n", "mean", "sd", "p0"))
   expect_identical(tail(input$stat), c(rep("top_counts", 5), rep("ordered", 1)))
   expect_identical(head(input$level), rep(".all", 6))
   expect_identical(tail(input$level),
-                   c("linseed", "sunflower", "meatmeal", 
+                   c("linseed", "sunflower", "meatmeal",
                      "horsebean", NA, ".all"  ))
   expect_equal(head(input$value), c(0, 71, 71, 261.3, 78.1, 108), tol = .01)
   expect_equal(tail(input$value), c( 12, 12, 11, 10, 0, 0))
@@ -46,18 +46,18 @@ test_that("Using skim_tee prints out the object", {
 
 test_that("Skimming a grouped data frame works as expected", {
   input <- dplyr::group_by(mtcars, cyl, gear) %>% skim()
-  
+
   # dimensions
   expect_length(input, 8)
   expect_equal(nrow(input), 792)
-  
+
   # classes
   expect_is(input, "skim_df")
   expect_is(input, "grouped_df")
   expect_is(input, "tbl_df")
   expect_is(input, "tbl")
   expect_is(input, "data.frame")
-  
+
   # values
   expect_identical(input$cyl, rep(c(4, 6, 8), c(297, 297, 198)))
   expect_true(all(c(3, 4, 5) %in% input$gear))
@@ -79,10 +79,10 @@ test_that("skim_to_wide works as expected.", {
   expect_length(input, 16)
   expect_equal(nrow(input), 5)
   expect_identical(input$type, c("factor", rep("numeric", each = 4)))
-  expect_identical(input$variable, 
+  expect_identical(input$variable,
     c("Species", "Petal.Length", "Petal.Width", "Sepal.Length", "Sepal.Width"))
   expect_identical(input$n, rep("150", each = 5))
-  expect_identical(input$top_counts, c("set: 50, ver: 50, vir: 50, NA: 0", 
+  expect_identical(input$top_counts, c("set: 50, ver: 50, vir: 50, NA: 0",
                                        NA, NA, NA, NA))
 })
 
@@ -93,7 +93,7 @@ test_that("skim_to_list works as expected", {
   expect_identical(class(input), "list")
   expect_identical(class(input[["numeric"]]), c("tbl", "tbl_df", "data.frame"))
   expect_equal(dim(input[["numeric"]]), c(1, 12))
-  expect_identical(names(input[["numeric"]]), 
+  expect_identical(names(input[["numeric"]]),
                    c("variable", "missing", "complete", "n", "mean",
                      "sd", "p0", "p25", "median", "p75", "p100", "hist" ))
 })
@@ -103,7 +103,7 @@ test_that("skim_to_list works with grouped data", {
   input <- skim_to_list(xg)
   expect_length(input,1)
[…]
