### 2018-02-14 fbe936c
Switch to github_document in Rmd
