+++ README.md
@@ -1,336 +1,275 @@
----
-output: 
-  html_document: 
-    keep_md: yes
----
-# skimr
 
+skimr
+=====
 
-[![Build Status](https://travis-ci.org/ropenscilabs/skimr.svg?branch=master)](https://travis-ci.org/ropenscilabs/skimr)
-[![codecov](https://codecov.io/gh/ropenscilabs/skimr/branch/master/graph/badge.svg)](https://codecov.io/gh/ropenscilabs/skimr)
-[![](https://badges.ropensci.org/175_status.svg)](https://github.com/ropensci/onboarding/issues/175)
+[![Build Status](https://travis-ci.org/ropenscilabs/skimr.svg?branch=master)](https://travis-ci.org/ropenscilabs/skimr) [![codecov](https://codecov.io/gh/ropenscilabs/skimr/branch/master/graph/badge.svg)](https://codecov.io/gh/ropenscilabs/skimr) [![](https://badges.ropensci.org/175_status.svg)](https://github.com/ropensci/onboarding/issues/175)
 
-The goal of skimr is to provide a frictionless approach to dealing with 
-summary statistics iteratively and interactively as part of a pipeline, 
-and that conforms to the principle of least surprise. 
+The goal of skimr is to provide a frictionless approach to dealing with summary statistics iteratively and interactively as part of a pipeline, and that conforms to the principle of least surprise.
 
-`Skimr` provides summary statistics that you can skim quickly to understand 
-your data and see what may be missing. It handles different data types 
-(numerics, factors, etc), and returns a skimr object that can be piped or 
-displayed nicely for the human reader. 
+`Skimr` provides summary statistics that you can skim quickly to understand your data and see what may be missing. It handles different data types (numerics, factors, etc), and returns a skimr object that can be piped or displayed nicely for the human reader.
 
-## Installation
+Installation
+------------
 
-The current released version of Skimr can be install from CRAN. 
-If you wish to install the current build of the next release you can do so 
-using the following:
+The current released version of Skimr can be install from CRAN. If you wish to install the current build of the next release you can do so using the following:
 
 ``` r
 # install.packages("devtools")
 devtools::install_github("ropenscilabs/skimr")
 ```
-The APIs for this branch should be considered reasonably stable 
-but still subject to change if an issue is discovered.
 
-To install the version with the most recent changes that have not yet been 
-incorporated in the master branch (and may not be):
+The APIs for this branch should be considered reasonably stable but still subject to change if an issue is discovered.
+
+To install the version with the most recent changes that have not yet been incorporated in the master branch (and may not be):
 
 ``` r
 devtools::install_github("ropenscilabs/skimr", ref = "develop")
 ```
+
 Do not rely on APIs from the develop branch.
 
-## Skim statistics in the console
+Skim statistics in the console
+------------------------------
 
[…]
