### 2018-02-17 #175 @elinw
I have a question for @jimhester, I did try moving the Rmd out of the docs folder to the top level (it's that way in develop right now)  but I really feel it makes the top level a bit messy there since it doesn't really serve any purpose for github users to have it there (unlike all the rd files).
