### 2018-02-17 ropensci/skimr#286 @elinw
Onboarding Issues: Readme
I'm going to use this issue to keep a record of responses to the onboarding comments about the README.md document.
This has been handled in a number of separate issues and commits.
Please edit the top to add more information or updates, add comments to comment.
Master history: https://github.com/ropenscilabs/skimr/commits/develop/README.md
Current diff of develop and master README.
https://github.com/ropenscilabs/skimr/pull/285/files#diff-5140cd39fa2897e01ffbfb8db0999940

- [x] Clean up extraneous text, yaml, and correct typos:
 - [x]  Change content:
 - [x] Move Rmd to top level folder: Done but uncertain about
- [x] Add links (to pander, kable, least surprise): Partly done, question asked.
 - [x] Improve writing:
