### 2018-02-17 5d2018a
Add references to Pillar, fix group example man, fix  max_chars example man.
