### 2018-02-17 #175 @elinw
Okay I'm starting a list of issues from  onboarding comments. I'll keep adding them here until I get a full list.
- https://github.com/ropenscilabs/skimr/issues/286  is for everything related to README
