### 2018-02-17 ropenscilabs/skimr#286 @elinw
@jimhester Can you clarify where you think links to pander and kable should go? Should they go to the pander and knitr packages? Specific functions in those packages? Or to the documentation of our implementation?
