### 2018-02-18 eae29a7
Examples of skim_tee(), skim_to_wide(), and skim_to_list().
