+++ man/skim.Rd
@@ -33,8 +33,10 @@ Printing does not change the structure of the \code{skim_df}, which remains a lo
 tibble.
 
 If you just want to see the printed output, call \code{skim_tee()} instead.
-This function returns the original data frame. If you want to work with
-a data frame that resembles the printed output, call \code{\link[=skim_to_wide]{skim_to_wide()}} or
+This function returns the original data frame.
+
+If you want to work with a data frame that resembles the printed output,
+call \code{\link[=skim_to_wide]{skim_to_wide()}} or for a named list of data frames by type
 \code{\link[=skim_to_list]{skim_to_list()}}. Note that all of the columns in the data frames produced
 by these functions are character. The intent is that you will be processing
 the \strong{printed} result further, not the original data.
@@ -79,4 +81,8 @@ skim(iris) \%>\% dplyr::filter(type == "factor")
 skim(iris) \%>\%
   dplyr::filter(stat == "mean") \%>\%
   dplyr::arrange(dplyr::desc(value))
+
+# Use skim_tee to view the skim results and
+# continue using the original data.
+chickwts \%>\% skim_tee() \%>\% dplyr::filter(feed == "sunflower")
 }
