### 2018-02-21 e758d5c
Treat Inf as NA for inline histogram
- Following the changes @elinw suggested!
- Change order of code so any mixture of only NA and Infs will give a
  blank histogram.
- Add elinw's test for correctly calculating inline histogram
