### 2018-02-25 0961e27
Add examples of pander and kable to readme and some other small updates.
