+++ README.md
@@ -1,3 +1,4 @@
+<!-- README.md is generated from README.Rmd. Please edit that file -->
 skimr
 =====
 
@@ -41,12 +42,12 @@ Skim statistics in the console
 
 -   Provides a larger set of statistics than `summary()`, including
     missing, complete, n, sd
--   reports numeric/int/dbl types separately from factor/chr
--   handles dates, logicals
+-   reports each data types separately
+-   handles dates, logicals, and a variety of other types
 -   supports spark-bar and spark-line based on [Hadley Wickham's pillar
     package](https://github.com/hadley/pillar).
 
-**Nicely separates variables by class:**
+### Nicely separates variables by class:
 
     skim(chickwts)
 
@@ -62,7 +63,7 @@ Skim statistics in the console
     ##  variable missing complete  n   mean    sd  p0   p25 median   p75 p100     hist
     ##    weight       0       71 71 261.31 78.07 108 204.5    258 323.5  423 ▃▅▅▇▃▇▂▂
 
-**Presentation is in a compact horizontal format:**
+### Presentation is in a compact horizontal format:
 
     skim(iris)
 
@@ -81,8 +82,7 @@ Skim statistics in the console
     ##  Sepal.Length       0      150 150 5.84 0.83 4.3 5.1   5.8  6.4  7.9 ▂▇▅▇▆▅▂▂
     ##   Sepal.Width       0      150 150 3.06 0.44 2   2.8   3    3.3  4.4 ▁▂▅▇▃▂▁▁
 
-**Individual columns of a data frame can be selected using
-tidyverse-style selectors:**
+### Individual columns can be selected using tidyverse-style selectors:
 
     skim(iris, Sepal.Length, Petal.Length)
 
@@ -95,7 +95,7 @@ tidyverse-style selectors:**
     ##  Petal.Length       0      150 150 3.76 1.77 1   1.6   4.35 5.1  6.9 ▇▁▁▂▅▅▃▁
     ##  Sepal.Length       0      150 150 5.84 0.83 4.3 5.1   5.8  6.4  7.9 ▂▇▅▇▆▅▂▂
 
-**Handles grouped data: **
+### Handles grouped data:
 
 `skim()` can handle data that has been grouped using `dplyr::group_by`.
 
@@ -121,10 +121,203 @@ tidyverse-style selectors:**
     ##   virginica Sepal.Length       0       50 50 6.59 0.64 4.9 6.23   6.5  6.9   7.9 ▁▁▃▇▅▃▂▃
     ##   virginica  Sepal.Width       0       50 50 2.97 0.32 2.2 2.8    3    3.18  3.8 ▁▃▇▇▅▃▁▂
 
-**Options for kable and pander: **
+### Options for kable and pander:
 
 Enhanced print options are available by piping to `kable()` or
-`pander()`.
+`pander()`. These build on the [pander
+package](https://cran.r-project.org/web/packages/pander/index.html) and
+the kable function of the [knitr
+package](https://cran.r-project.org/web/packages/knitr/index.html) These
+examples show how the enhanced options appear in the console. Note that
+the `skimr::` names space is used to prevent it being replaced by
+knitr::kable (which will result in the long skim\_df object being
+printed.)
+
+    skim(iris) %>% skimr::kable()
+
+    ## Skim summary statistics  
+    ##  n obs: 150    
+    ##  n variables: 5    
+    ## 
+    ## Variable type: factor
+    ## 
+    ## variable   missing   complete   n     n_unique   top_counts                         ordered 
[…]
