### 2018-02-25 763df94
Fix kable rendering. Improve organization.
