+++ README.md
@@ -41,13 +41,13 @@ Skim statistics in the console
 `skimr`:
 
 -   Provides a larger set of statistics than `summary()`, including
-    missing, complete, n, sd
+    missing, complete, n, and sd.
 -   reports each data types separately
 -   handles dates, logicals, and a variety of other types
 -   supports spark-bar and spark-line based on [Hadley Wickham's pillar
     package](https://github.com/hadley/pillar).
 
-### Nicely separates variables by class:
+### Separates variables by class:
 
     skim(chickwts)
 
@@ -81,6 +81,53 @@ Skim statistics in the console
     ##   Petal.Width       0      150 150 1.2  0.76 0.1 0.3   1.3  1.8  2.5 ▇▁▁▅▃▃▂▂
     ##  Sepal.Length       0      150 150 5.84 0.83 4.3 5.1   5.8  6.4  7.9 ▂▇▅▇▆▅▂▂
     ##   Sepal.Width       0      150 150 3.06 0.44 2   2.8   3    3.3  4.4 ▁▂▅▇▃▂▁▁
+
+### Built in support for strings, lists and other column classes
+
+    skim(dplyr::starwars)
+
+    ## Skim summary statistics
+    ##  n obs: 87 
+    ##  n variables: 13 
+    ## 
+    ## Variable type: character 
+    ##    variable missing complete  n min max empty n_unique
+    ##   eye_color       0       87 87   3  13     0       15
+    ##      gender       3       84 87   4  13     0        4
+    ##  hair_color       5       82 87   4  13     0       12
+    ##   homeworld      10       77 87   4  14     0       48
+    ##        name       0       87 87   3  21     0       87
+    ##  skin_color       0       87 87   3  19     0       31
+    ##     species       5       82 87   3  14     0       37
+    ## 
+    ## Variable type: integer 
+    ##  variable missing complete  n   mean    sd p0 p25 median p75 p100     hist
+    ##    height       6       81 87 174.36 34.77 66 167    180 191  264 ▁▁▁▂▇▃▁▁
+    ## 
+    ## Variable type: list 
+    ##   variable missing complete  n n_unique min_length median_length max_length
+    ##      films       0       87 87       24          1             1          7
+    ##  starships       0       87 87       17          0             0          5
+    ##   vehicles       0       87 87       11          0             0          2
+    ## 
+    ## Variable type: numeric 
+    ##    variable missing complete  n  mean     sd p0  p25 median  p75 p100     hist
+    ##  birth_year      44       43 87 87.57 154.69  8 35       52 72    896 ▇▁▁▁▁▁▁▁
+    ##        mass      28       59 87 97.31 169.46 15 55.6     79 84.5 1358 ▇▁▁▁▁▁▁▁
+
+### Has a useful summary function
+
+    skim(iris) %>% summary()
+
+    ## A skim object    
+    ## 
+    ## Name: iris   
+    ## Number of Rows: 150   
+    ## Number of Columns: 5    
+    ##     
+    ## Column type frequency    
+    ## factor: 1   
+    ## numeric: 4
 
 ### Individual columns can be selected using tidyverse-style selectors:
 
@@ -128,31 +175,123 @@ Enhanced print options are available by piping to `kable()` or
 package](https://cran.r-project.org/web/packages/pander/index.html) and
 the kable function of the [knitr
 package](https://cran.r-project.org/web/packages/knitr/index.html) These
-examples show how the enhanced options appear in the console. Note that
[…]
