### 2018-02-26 8991896
More editing of documentation. Add vignette about defaults.
