+++ README.Rmd
@@ -78,13 +78,13 @@ skim(iris) %>% summary()
 ```
 
 
-### Individual columns can be selected using tidyverse-style selectors:
+### Individual columns can be selected using tidyverse-style selectors
 
 ```{r}
 skim(iris, Sepal.Length, Petal.Length)
 ```
 
-### Handles grouped data: 
+### Handles grouped data 
 
 `skim()` can handle data that has been grouped using `dplyr::group_by`.
 
@@ -92,20 +92,32 @@ skim(iris, Sepal.Length, Petal.Length)
 iris %>% dplyr::group_by(Species) %>% skim()
 ```
 
-### Options for kable and pander:
+## Knitted results
+
+Simply skimming a data frame will produce the horizontal print
+layout shown above. When knitting you can also used enhanced rendering
+with kable and pander implementations. 
+
+### Options for kable and pander
 
 Enhanced print options are available by piping to `kable()` or `pander()`.
 These build on the [pander package](https://cran.r-project.org/web/packages/pander/index.html)
 and the kable function of the [knitr package](https://cran.r-project.org/web/packages/knitr/index.html)
 These examples show how the enhanced options should appear after 
 knitting, however your results may differ (see vignettes for details).
-Note that the `skimr::` names space is used to prevent it being
+
+
+### Option for kable. 
+Note that the results='asis' chunk option is used and  the
+`skimr::` namespace is used to prevent it being
 replaced by knitr::kable (which will result in the long skim_df 
 object being printed.)
-
 ```{r results='asis'}
 skim(iris) %>% skimr::kable()
 ```
+### Options for pander
+
+At times you may need `panderOptions('knitr.auto.asis', FALSE)`.
 
 ```{r}
 skim(iris) %>% pander()
@@ -126,14 +138,23 @@ dim(a)
 print.data.frame(skim(chickwts))
 ```
 
-## Compute on the full `skim_df` object
+### Compute on the full `skim_df` object
 ```{r}
 skim(mtcars) %>% dplyr::filter(stat=="hist")
 ```
 
-Users can add new classes.
+## Customizing skimr
 
-## Specify your own statistics
+Although skimr provides opinionated defaults, it is highly customizable.
+Users can specify their own statistics, change the formatting of results,
+create statistics for new classes and develop skimmers for data structures
+that are not data frames.
+
+### Specify your own statistics and classes
+
+Users can specify their own statistics using a list combined with the
+`skim_with()` function.  This can support any named class found in
+your data.
 
 ```{r}
 funs <- list(
@@ -149,6 +170,18 @@ skim_with_defaults()
 
 ```
 
+### Change formatting
+
+Skimr provides a set of default formats that allow decimals in columns to
+be aligned, a reasonable number of decimal places for numeric data, and a 
+representation of dates.  Users can view thes with `show_formats()` and
+modify them with `skim_format()`.
+
+### Skimming other objects
+
+Procedures for developing skim functions for other objects are described in 
+the vignette _Supporting additional objects_.
+
 ## Limitations of current version
 
 We are aware that there are issues with rendering the inline histograms and
@@ -159,8 +192,9 @@ line charts in various contexts, some of which are described below.
 There are known issues with printing the spark-histogram characters when
 printing a data frame. For example, `"▂▅▇"` is printed as
 `"<U+2582><U+2585><U+2587>"`. This longstanding problem [originates in
-the low-level code](http://r.789695.n4.nabble.com/Unicode-display-problem-with-data-frames-under-Windows-td4707639.html) for printing dataframes; for example,
-there are reports of this issue in Emacs ESS.
[…]
