### 2018-02-26 76ba7c7
Fix  bad HTML from kable.
