### 2018-02-26 ropenscilabs/skimr#286 @elinw
Okay at this point I am pretty sure we've addressed the comments about the read me. I think it reads much better now.
