### 2018-02-26 8f11bd5
Update news and description files.
