+++ DESCRIPTION
@@ -14,7 +14,12 @@ Authors@R: c(
   person("Kirill", "Müller", email="krlmlr+r@mailbox.org", role = 'ctb'),
   person(family = "RStudio, Inc.", role = "cph", comment = "Spark functions"),
   person("Connor", "Kirkpatrick", email="hello@connorkirkpatrick.com", role = 'ctb'),
-  person("Scott","Brenstuhl", email="brenstsr@miamioh.edu", role = 'ctb')
+  person("Scott", "Brenstuhl", email="brenstsr@miamioh.edu", role = 'ctb'),
+  person("Patrick", "Schratz", email="patrick.schratz@gmail.com", role = 'ctb'),
+  person("lbusett", "", email="lbusett@gmail.com", role='ctb'),
+  person("Mikko","Korpela", email="mvkorpel@iki.fi", role='ctb'),
+  person("Jennifer","Thompson", email="thompson.jennifer@gmail.com", role='ctb'),
+  person("Harris","McGehee", email="mcgehee.harris@gmail.com", role='ctb')
   )
 Description: A simple to use summary function that can be used with pipes
     and displays nicely in the console. The default summary statistics may be 
