### 2018-02-26 cc07dbb
Update windows tests.
