### 2018-03-03 #175 @elinw
An update.

As of today I think the easy 3 (that wouldn't involve API changes) have all been addressed.  @michaelquinn32  has been working on a discussion paper about the API issues so that we can have a clear plan moving toward skimr 2.0 and our user community can chime in.   That should be ready to share sson.
