### 2018-03-05 c4eea84
Fix error when a factor has a 0 length label (which is invalid but can happen as shown in issue report).  Closes 299.
