### 2018-03-07 3b5ad2b
Move tests to the right files and add a warning to inline_hist() when Inf values are present.
