### 2018-03-23 10d8978
Set minimum version of knitr to deal with 64 bit rendering problem. Issue #297
