### 2018-03-23 73b94f1
Fix typo and add news item.
