### 2018-03-29 361c4d0
Don't attempt to display a data frame name when it is unavailable because of piping,
