+++ NEWS.md
@@ -19,6 +19,8 @@ skimr 1.0.2 (2018-xx-xx)
     `skim_with(iqr = IQR, hist = NULL)` works as expected.
   * Histograms work when Inf or -Inf are present.
   * Change seq( ) parameter to length.out to avoid problems with name matching.
+  * Summary should not display a data frame name of "." 
+    (which occurs when piping begins with the data frame).
 
 skimr 1.0.1 (2018-01-09)
 ========================
