+++ README.md
@@ -60,8 +60,8 @@ Skim statistics in the console
     ##      feed       0       71 71        6 soy: 14, cas: 12, lin: 12, sun: 12   FALSE
     ## 
     ## Variable type: numeric 
-    ##  variable missing complete  n   mean    sd  p0   p25 median   p75 p100     hist
-    ##    weight       0       71 71 261.31 78.07 108 204.5    258 323.5  423 ▃▅▅▇▃▇▂▂
+    ##  variable missing complete  n   mean    sd  p0   p25 p50   p75 p100     hist
+    ##    weight       0       71 71 261.31 78.07 108 204.5 258 323.5  423 ▃▅▅▇▃▇▂▂
 
 ### Presentation is in a compact horizontal format:
 
@@ -76,11 +76,11 @@ Skim statistics in the console
     ##   Species       0      150 150        3 set: 50, ver: 50, vir: 50, NA: 0   FALSE
     ## 
     ## Variable type: numeric 
-    ##      variable missing complete   n mean   sd  p0 p25 median p75 p100     hist
-    ##  Petal.Length       0      150 150 3.76 1.77 1   1.6   4.35 5.1  6.9 ▇▁▁▂▅▅▃▁
-    ##   Petal.Width       0      150 150 1.2  0.76 0.1 0.3   1.3  1.8  2.5 ▇▁▁▅▃▃▂▂
-    ##  Sepal.Length       0      150 150 5.84 0.83 4.3 5.1   5.8  6.4  7.9 ▂▇▅▇▆▅▂▂
-    ##   Sepal.Width       0      150 150 3.06 0.44 2   2.8   3    3.3  4.4 ▁▂▅▇▃▂▁▁
+    ##      variable missing complete   n mean   sd  p0 p25  p50 p75 p100     hist
+    ##  Petal.Length       0      150 150 3.76 1.77 1   1.6 4.35 5.1  6.9 ▇▁▁▂▅▅▃▁
+    ##   Petal.Width       0      150 150 1.2  0.76 0.1 0.3 1.3  1.8  2.5 ▇▁▁▅▃▃▂▂
+    ##  Sepal.Length       0      150 150 5.84 0.83 4.3 5.1 5.8  6.4  7.9 ▂▇▅▇▆▅▂▂
+    ##   Sepal.Width       0      150 150 3.06 0.44 2   2.8 3    3.3  4.4 ▁▂▅▇▃▂▁▁
 
 ### Built in support for strings, lists and other column classes
 
@@ -101,8 +101,8 @@ Skim statistics in the console
     ##     species       5       82 87   3  14     0       37
     ## 
     ## Variable type: integer 
-    ##  variable missing complete  n   mean    sd p0 p25 median p75 p100     hist
-    ##    height       6       81 87 174.36 34.77 66 167    180 191  264 ▁▁▁▂▇▃▁▁
+    ##  variable missing complete  n   mean    sd p0 p25 p50 p75 p100     hist
+    ##    height       6       81 87 174.36 34.77 66 167 180 191  264 ▁▁▁▂▇▃▁▁
     ## 
     ## Variable type: list 
     ##   variable missing complete  n n_unique min_length median_length max_length
@@ -111,9 +111,9 @@ Skim statistics in the console
     ##   vehicles       0       87 87       11          0             0          2
     ## 
     ## Variable type: numeric 
-    ##    variable missing complete  n  mean     sd p0  p25 median  p75 p100     hist
-    ##  birth_year      44       43 87 87.57 154.69  8 35       52 72    896 ▇▁▁▁▁▁▁▁
[…]
