### 2018-04-02 #175 @elinw
Okay so we released 1.0.2 with all the bug and documentation fixes so far.  I'm going to look at this possibility of using  markup to make the display more clear to people.
