### 2018-04-03 29a52fd
Switch knitr import to a different function.
