+++ NAMESPACE
@@ -49,6 +49,7 @@ export(sorted_count)
 export(starts_with)
 export(ts_end)
 export(ts_start)
+importFrom(knitr,image_uri)
 importFrom(magrittr,"%>%")
 importFrom(pander,pander)
 importFrom(tidyselect,contains)
