### 2018-04-03 16ab7b9
Address an issue in which the knit::kable() was used when skimr::kable() was expected and a similar potential problem with pander().   Adjust tests for the outputs due to change in default alignments in kable().
