+++ NAMESPACE
@@ -3,8 +3,6 @@
 S3method(kable,data.frame)
 S3method(kable,skim_df)
 S3method(kable,summary_skim_df)
-S3method(pander,skim_df)
-S3method(pander,summary_skim_df)
 S3method(print,skim_df)
 S3method(print,skim_vector)
 S3method(print,spark)
@@ -35,6 +33,9 @@ export(n_missing)
 export(n_unique)
 export(num_range)
 export(one_of)
+export(pander)
+export(pander.skim_df)
+export(pander.summary_skim_df)
 export(show_formats)
 export(show_skimmers)
 export(skim)
@@ -49,9 +50,7 @@ export(sorted_count)
 export(starts_with)
 export(ts_end)
 export(ts_start)
-importFrom(knitr,image_uri)
 importFrom(magrittr,"%>%")
-importFrom(pander,pander)
 importFrom(tidyselect,contains)
 importFrom(tidyselect,ends_with)
 importFrom(tidyselect,everything)
