### 2018-04-03 2dc4d3b
Fix Windows only test.
