### 2018-04-03 6043d70
Fix expected results on 32 bit test.
