### 2018-04-04 6935b3b
Fix spaces in vector output.
This is probably due to a change in labels.
