### 2018-04-04 41a6321
Simplify calls to rlang::sym.
