### 2018-04-04 ea198d1
Update test-skim_print.R
