### 2018-04-06 4b3b8c0
Skip skim_v tests on windows
