### 2018-04-06 8d0e2ab
Add a helper for fixing histograms on windows.
This wraps the locale change with a prompt.
