+++ R/utils.R
@@ -22,3 +22,23 @@ make_utf8 <- function(x) {
 is_windows <- function() {
   tolower(Sys.info()[["sysname"]]) == "windows"
 }
+
+#' Fix unicode histograms on Windows
+#' 
+#' This functions changes your session's locale to address issues with printing
+#' histograms on Windows.
+#' 
+#' There are known issues with printing the spark-histogram characters when
+#' printing a data frame. For example, "▂▅▇" is printed as
+#' "<U+2582><U+2585><U+2587>". This longstanding problem originates in the
+#' low-level code for printing dataframes.
+#' @export
+
+fix_windows_histograms <- function() {
+  message("This function will change your system locale. It may have other ",
+          "unintended effects.")
+  response <- readline("Continue? (Y/n)")
+  if (tolower(response) != "n") Sys.setlocale("LC_CTYPE", "Chinese")
+  else message("Locale was not changed.")
+  invisible(NULL)
+}
