+++ man/fix_windows_histograms.Rd
@@ -1,0 +1,18 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/utils.R
+\name{fix_windows_histograms}
+\alias{fix_windows_histograms}
+\title{Fix unicode histograms on Windows}
+\usage{
+fix_windows_histograms()
+}
+\description{
+This functions changes your session's locale to address issues with printing
+histograms on Windows.
+}
+\details{
+There are known issues with printing the spark-histogram characters when
+printing a data frame. For example, "▂▅▇" is printed as
+"<U+2582><U+2585><U+2587>". This longstanding problem originates in the
+low-level code for printing dataframes.
+}
