+++ NAMESPACE
@@ -15,6 +15,7 @@ export("%>%")
 export(contains)
 export(ends_with)
 export(everything)
+export(fix_windows_histograms)
 export(get_skimmers)
 export(inline_hist)
 export(inline_linegraph)
@@ -50,7 +51,6 @@ export(sorted_count)
 export(starts_with)
 export(ts_end)
 export(ts_start)
-importFrom(knitr,kable)
 importFrom(magrittr,"%>%")
 importFrom(tidyselect,contains)
 importFrom(tidyselect,ends_with)
