### 2018-04-06 2eb81f4
Remove unicode characters from docs.
