### 2018-04-06 43026ba
Add .list argument to options.
This lets you set multiple arguments that you can store in a list.
