+++ R/functions.R
@@ -16,6 +16,9 @@ NULL
 #' 
 #' @param ... A list of functions, with an argument name that matches a
 #'  particular data type.
+#' @param .list Instead of individual named entries, you can provided a list
+#'  instead. If most \code{...} arguments and \code{.list} is provided, values
+#'  in \code{.list} take precedence.
 #' @param append Whether the provided options should be in addition to the
 #'  defaults already in `skim`. Default is `TRUE`.
 #' @return When setting values, `invisible(NULL)`. When looking up values a
@@ -48,6 +51,14 @@ NULL
 #' # Or use an rlang-style formula constructor for the function
 #' skim_with(numeric = list(med = ~median(., na.rm = TRUE)))
 #' 
+#' # Set multiple types of skimmers simultaneously
+#' skim_with(numeric = list(mean = mean), character = list(len = length))
+#' 
+#' # Or pass the same as a list
+#' my_skimmers <- list(numeric = list(mean = mean),
+#'                    character = list(len = length))
+#' skim_with(.list = my_skimmers)
+#' 
 #' # Go back to defaults
 #' skim_with_defaults()
 #' skim(faithful)
@@ -61,8 +72,9 @@ NULL
 #' skim_with(hms = funs$date)
 #' @export
 
-skim_with <- function(..., append = TRUE, drop_new = FALSE) {
-  skimmers <- purrr::modify_depth(rlang::dots_list(...), 1, purrr::map_if,
+skim_with <- function(..., .list = list(), append = TRUE, drop_new = FALSE) {
+  combined <- purrr::list_modify(list(...), !!! .list)
+  skimmers <- purrr::modify_depth(combined, 1, purrr::map_if,
                                   Negate(is.null), rlang::as_function)
   skim_options(skimmers, env = "functions", append = append,
                drop_new = drop_new)
@@ -73,7 +85,7 @@ skim_with <- function(..., append = TRUE, drop_new = FALSE) {
 #' @export
 
 skim_with_defaults <- function() {
-  do.call(skim_with, c(.default, append = FALSE, drop_new = TRUE))
+  skim_with(.list = .default, append = FALSE, drop_new = TRUE)
 }
 
 
