+++ tests/testthat/test-formats.R
@@ -23,6 +23,22 @@ test_formats_that("Formatting options can be changed", {
   expect_identical(formats$integer, list(drop0trailing = TRUE))
 })
 
+test_formats_that("Multiple formats can be changed", {
+  skim_format(numeric = list(digits = 1), integer = list(digits = 1),
+              append = FALSE) 
+  formats <- show_formats()
+  expect_identical(formats$numeric, list(digits = 1))
+  expect_identical(formats$integer, list(digits = 1))
+})
+
+test_formats_that("Multiple formats can be changed, list", {
+  new_formats <- list(numeric = list(digits = 1), integer = list(digits = 1))
+  skim_format(.list = new_formats, append = FALSE) 
+  formats <- show_formats()
+  expect_identical(formats$numeric, list(digits = 1))
+  expect_identical(formats$integer, list(digits = 1))
+})
+
 test_formats_that("Formatting options change printed output", {
   skim_format(numeric = list(digits = 2, nsmall = 2), .align_decimal = FALSE,
               append = FALSE)
