+++ man/kable.Rd
@@ -17,40 +17,40 @@ kable(x, format = NULL, digits = getOption("digits"), row.names = NA,
 \method{kable}{summary_skim_df}(x, ...)
 }
 \arguments{
-\item{x}{An R object, typically a matrix or data frame.}
+\item{x}{an R object (typically a matrix or data frame)}
 
-\item{format}{A character string. Possible values are \code{latex},
+\item{format}{a character string; possible values are \code{latex},
 \code{html}, \code{markdown}, \code{pandoc}, and \code{rst}; this will be
 automatically determined if the function is called within \pkg{knitr}; it
-can also be set in the global option \code{knitr.table.format}. If
-\code{format} is a function, it must return a character string.}
+can also be set in the global option \code{knitr.table.format}; if
+\code{format} is a function, it must return a character string}
 
-\item{digits}{Maximum number of digits for numeric columns, passed to
-\code{round()}. This can also be a vector of length \code{ncol(x)}, to set
-the number of digits for individual columns.}
+\item{digits}{the maximum number of digits for numeric columns (passed to
+\code{round()}); it can also be a vector of length \code{ncol(x)} to set
+the number of digits for individual columns}
 
-\item{row.names}{Logical: whether to include row names. By
+\item{row.names}{a logical value indicating whether to include row names; by
 default, row names are included if \code{rownames(x)} is neither
-\code{NULL} nor identical to \code{1:nrow(x)}.}
+\code{NULL} nor identical to \code{1:nrow(x)}}
 
-\item{col.names}{A character vector of column names to be used in the table.}
+\item{col.names}{a character vector of column names to be used in the table}
 
-\item{align}{Column alignment: a character vector consisting of
- \code{'l'} (left), \code{'c'} (center) and/or \code{'r'} (right). By
- default or if \code{align = NULL}, numeric columns are right-aligned, and
- other columns are left-aligned. If \code{length(align) == 1L}, the string will be
- expanded to a vector of individual letters, e.g. \code{'clc'} becomes
-\code{c('c', 'l', 'c')}, unless the output format is LaTeX.}
+\item{align}{the alignment of columns: a character vector consisting of
+\code{'l'} (left), \code{'c'} (center) and/or \code{'r'} (right); by
+default, numeric columns are right-aligned, and other columns are
+left-aligned; if \code{align = NULL}, the default alignment is used;
+alternatively, if \code{length(align) == 1L}, the string will be expanded
+to a vector of individual letters unless the output format is LaTeX; for
+example, \code{'clc'} will be converted to \code{c('c', 'l', 'c')}}
 
-\item{caption}{The table caption.}
+\item{caption}{the table caption}
 
-\item{format.args}{A list of arguments to be passed to \code{\link{format}()}
-to format table values, e.g. \code{list(big.mark = ',')}.}
+\item{format.args}{a list of arguments to be passed to \code{\link{format}()}
+to format table values, e.g. \code{list(big.mark = ',')}}
 
-\item{escape}{Boolean; whether to escape special characters when producing
-HTML or LaTeX tables.}
+\item{escape}{escape special characters when producing HTML or LaTeX tables}
 
-\item{...}{Other arguments (see Examples).}
+\item{...}{other arguments (see examples)}
 }
 \value{
 The original \code{skim_df} object.
