+++ R/formats.R
@@ -34,7 +34,7 @@ NULL
 #' skim_format(numeric = list(digits = 3))
 #' 
 #' # Show the values for the formats
-#' show_formats
+#' show_formats()
 #' 
 #' # Show 4-character names in factor levels
 #' skim_format(.levels = list(max_char = 4))
@@ -43,15 +43,18 @@ NULL
 #' skim_format(numeric = list(digits = 3), .levels = list(max_char = 4))
 #' 
 #' # Set multiple formats with a .list argument
-#' myFormats <- list(numeric = list(digits = 3), .levels = list(max_char = 4))
-#' skim_format(.list = myFormats)
+#' my_formats <- list(numeric = list(digits = 3), .levels = list(max_char = 4))
+#' skim_format(.list = my_formats)
+#' 
+#' # Alternatively, use rlang unquoting semantics
+#' skim_format(!!!my_formats)
 #' 
 #' # Reset to the defaults
 #' skim_format_defaults()
 #' @export
 
 skim_format <- function(..., .list = list(), append = TRUE, drop_new = FALSE) {
-  combined <- purrr::list_modify(list(...), !!! .list)
+  combined <- purrr::list_modify(rlang::dots_list(...), !!! .list)
   skim_options(combined, env = "formats", append = append, drop_new = drop_new)
 }
 
