### 2018-04-06 80f65f0
Update news and add some help files that were missed.
