### 2018-04-08 47fea60
When using skim_with() an empty list for a class should explicitly eliminate the class from skimming (and should not fall back to character). When skimming a single vector of the excluded type, a message should be returned.
