### 2018-04-08 #175 @elinw
@maelle That's the plan.  I wanted to add that we now have in the develop branch a solution to the other issue that I think folks here and in our tracker have wanted.  That is now you can define your own complete list of functions with a name and use skim_with() to append or replace the default skimmers.  So the way I think about this the use case would be that you have several different lists that you use for different situations e.g. my_funs1, my_funs2, my_funs3.   Then you still have to do two steps but it is e.g.

```
skim_with(.list = my_funs1, append = FALSE)
skim(x)
skim_with(.list = my_funs2, append = FALSE)
skim(y)
skim_with(!!!my_funs3, append=FALSE)
skim_with_defaults()
```
I think if this were part of your usual workflow you might write your own functions so you don't need to do that `append=FALSE` and `.list=`.

There's also a PR making it so that if you define an empty list of functions for a type that type is removed from the analysis.
