### 2018-04-10 176bd20
Update manual files.
