### 2018-04-14 6f18791
Make the display of the default method more clearly show separation between types
