+++ README.md
@@ -54,12 +54,10 @@ Skim statistics in the console
     ## Skim summary statistics
     ##  n obs: 71 
     ##  n variables: 2 
-    ## 
-    ## Variable type: factor 
+    ## ── Variable type:factor ────────────────────────────────────────────
     ##  variable missing complete  n n_unique                         top_counts ordered
     ##      feed       0       71 71        6 soy: 14, cas: 12, lin: 12, sun: 12   FALSE
-    ## 
-    ## Variable type: numeric 
+    ## ── Variable type:numeric ───────────────────────────────────────────
     ##  variable missing complete  n   mean    sd  p0   p25 p50   p75 p100     hist
     ##    weight       0       71 71 261.31 78.07 108 204.5 258 323.5  423 ▃▅▅▇▃▇▂▂
 
@@ -70,12 +68,10 @@ Skim statistics in the console
     ## Skim summary statistics
     ##  n obs: 150 
     ##  n variables: 5 
-    ## 
-    ## Variable type: factor 
+    ## ── Variable type:factor ────────────────────────────────────────────
     ##  variable missing complete   n n_unique                       top_counts ordered
     ##   Species       0      150 150        3 set: 50, ver: 50, vir: 50, NA: 0   FALSE
-    ## 
-    ## Variable type: numeric 
+    ## ── Variable type:numeric ───────────────────────────────────────────
     ##      variable missing complete   n mean   sd  p0 p25  p50 p75 p100     hist
     ##  Petal.Length       0      150 150 3.76 1.77 1   1.6 4.35 5.1  6.9 ▇▁▁▂▅▅▃▁
     ##   Petal.Width       0      150 150 1.2  0.76 0.1 0.3 1.3  1.8  2.5 ▇▁▁▅▃▃▂▂
@@ -89,8 +85,7 @@ Skim statistics in the console
     ## Skim summary statistics
     ##  n obs: 87 
     ##  n variables: 13 
-    ## 
-    ## Variable type: character 
+    ## ── Variable type:character ─────────────────────────────────────────
     ##    variable missing complete  n min max empty n_unique
     ##   eye_color       0       87 87   3  13     0       15
     ##      gender       3       84 87   4  13     0        4
@@ -99,18 +94,15 @@ Skim statistics in the console
     ##        name       0       87 87   3  21     0       87
     ##  skin_color       0       87 87   3  19     0       31
     ##     species       5       82 87   3  14     0       37
-    ## 
-    ## Variable type: integer 
+    ## ── Variable type:integer ───────────────────────────────────────────
     ##  variable missing complete  n   mean    sd p0 p25 p50 p75 p100     hist
     ##    height       6       81 87 174.36 34.77 66 167 180 191  264 ▁▁▁▂▇▃▁▁
-    ## 
-    ## Variable type: list 
+    ## ── Variable type:list ──────────────────────────────────────────────
     ##   variable missing complete  n n_unique min_length median_length max_length
     ##      films       0       87 87       24          1             1          7
     ##  starships       0       87 87       17          0             0          5
     ##   vehicles       0       87 87       11          0             0          2
-    ## 
-    ## Variable type: numeric 
+    ## ── Variable type:numeric ───────────────────────────────────────────
[…]
