### 2018-05-20 1cde3ff
Remove deprecated functions.
