+++ DESCRIPTION
@@ -35,7 +35,6 @@ Imports:
     cli,
     dplyr (>= 0.7),
     magrittr,
-    pander,
     purrr,
     rlang,
     stats,
@@ -57,18 +56,15 @@ URL: https://github.com/ropenscilabs/skimr
 BugReports: https://github.com/ropenscilabs/skimr/issues
 VignetteBuilder: knitr
 Roxygen: list(markdown = TRUE)
-RoxygenNote: 6.0.1
 Collate: 
     'skimr-package.R'
     'formats.R'
     'stats.R'
-    'functions.R'
+    'get_skimmers.R'
     'kable.R'
-    'options.R'
-    'pander.R'
     'skim.R'
     'skim_print.R'
-    'skim_v.R'
     'summary.R'
     'utils.R'
     'wide.R'
+RoxygenNote: 6.0.1
