+++ NAMESPACE
@@ -1,5 +1,16 @@
 # Generated by roxygen2: do not edit by hand
 
+S3method(get_skimmers,AsIs)
+S3method(get_skimmers,Date)
+S3method(get_skimmers,POSIXct)
+S3method(get_skimmers,character)
+S3method(get_skimmers,complex)
+S3method(get_skimmers,default)
+S3method(get_skimmers,double)
+S3method(get_skimmers,factor)
+S3method(get_skimmers,integer)
+S3method(get_skimmers,logical)
+S3method(get_skimmers,ts)
 S3method(kable,data.frame)
 S3method(kable,skim_df)
 S3method(kable,summary_skim_df)
@@ -7,9 +18,6 @@ S3method(print,skim_df)
 S3method(print,skim_vector)
 S3method(print,spark)
 S3method(print,summary_skim_df)
-S3method(skim,data.frame)
-S3method(skim,default)
-S3method(skim,grouped_df)
 S3method(summary,skim_df)
 export("%>%")
 export(contains)
@@ -25,6 +33,7 @@ export(list_lengths_median)
 export(list_lengths_min)
 export(list_max_length)
 export(list_min_length)
+export(make_skimmer)
 export(matches)
 export(max_char)
 export(min_char)
@@ -34,19 +43,17 @@ export(n_missing)
 export(n_unique)
 export(num_range)
 export(one_of)
-export(pander)
-export(pander.skim_df)
-export(pander.summary_skim_df)
 export(show_formats)
-export(show_skimmers)
 export(skim)
+export(skim.data.frame)
+export(skim.default)
+export(skim.grouped_df)
 export(skim_format)
 export(skim_format_defaults)
 export(skim_tee)
 export(skim_to_list)
 export(skim_to_wide)
 export(skim_with)
-export(skim_with_defaults)
 export(sorted_count)
 export(starts_with)
 export(ts_end)
