### 2018-05-22 #175 @elinw
Hi,

Yes I think the development branch pretty much has addressed all, I just wanted to put in Jenny’s suggestion to make the display a bit nicer.  I’ll do that once I recover from the unconfirmed! And then I think we could call it ready and everything else in terms of possibly changing the API etc will be for 2.0.

> On May 22, 2018, at 1:32 PM, Maëlle Salmon <notifications@github.com> wrote:
>
> @elinw <https://github.com/elinw> @michaelquinn32 <https://github.com/michaelquinn32> any update? ☺
>
