### 2018-05-23 9c42cf6
Add a new interface for looking up skim functions.
