### 2018-05-23 ba2f957
Remove dependencies on global options and change functions to support new wide skimming approach.
