### 2018-05-23 3b1a51b
Create new functional approach to defining skimming functions.
