+++ man/skim_with.Rd
@@ -1,35 +1,19 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/functions.R
+% Please edit documentation in R/skim.R
 \name{skim_with}
 \alias{skim_with}
-\alias{skim_with_defaults}
-\alias{show_skimmers}
-\alias{get_skimmers}
 \title{Set or add the summary functions for a particular type of data}
 \usage{
-skim_with(..., .list = list(), append = TRUE, drop_new = FALSE)
-
-skim_with_defaults()
-
-show_skimmers(which = NULL)
-
-get_skimmers(which = NULL)
+skim_with(..., append = TRUE)
 }
 \arguments{
 \item{...}{A list of functions, with an argument name that matches a
 particular data type.}
 
-\item{.list}{Instead of individual named entries, you can provided a list
-instead. If most \code{...} arguments and \code{.list} is provided, values
-in \code{.list} take precedence.}
-
 \item{append}{Whether the provided options should be in addition to the
 defaults already in \code{skim}. Default is \code{TRUE}.}
 
 \item{drop_new}{Whether types outside of the defaults should be discarded.}
-
-\item{which}{A character vector. One or more of the classes whose summary
-functions you wish to display.}
 }
 \value{
 When setting values, \code{invisible(NULL)}. When looking up values a
@@ -51,16 +35,6 @@ This is an intentional design choice, with effects similar to setting
 options in base R. By setting options here for your entire session, you
 can continue to summarize using skim on its own.
 }
-\section{Functions}{
-\itemize{
-\item \code{skim_with_defaults}: Use the default functions within skim
-
-\item \code{show_skimmers}: Access the names of the summary functions for a
-class.
-
-\item \code{get_skimmers}: Pull a list of summary functions for a class.
-}}
-
 \examples{
 # Use new functions for numeric functions
 skim_with(numeric = list(median = median, mad = mad), append = FALSE)
@@ -91,9 +65,6 @@ skim_with(numeric = list(mean = mean), character = list(len = length))
 # Or pass the same as a list
 my_skimmers <- list(numeric = list(mean = mean),
                     character = list(len = length))
-skim_with(.list = my_skimmers)
-
-# Alternatively, use rlang unquoting semantics
 skim_with(!!!my_skimmers)
 
 # Go back to defaults
