### 2018-05-23 72fdd0c
Remove global options.
