+++ R/skim.R
@@ -1,5 +1,140 @@
 globalVariables(".")
 
+#' Set or add the summary functions for a particular type of data
+#' 
+#' While skim is designed around having an opinionated set of defaults, you
+#' can use these functions to change the summary statistics that it returns.
+#' To do that, provide type you wish to change as an argument to this function,
+#' along with a list of named functions that you want to use instead of the
+#' defaults. 
+#' 
+#' `skim_with()` is a closure: a function that returns a new function. This
+#' lets you have several skimming functions in a single R session, but it
+#' also means that you need to assign the return of `skim_with()` before
+#' you can use it.
+#' 
+#' @param ... A list of functions, with an argument name that matches a
+#'  particular data type.
+#' @param append Whether the provided options should be in addition to the
+#'  defaults already in `skim`. Default is `TRUE`.
+#' @return A new `skim()` function. This is callable. See [skim()] for
+#'  more details.
+#' @examples
+#' # Use new functions for numeric functions
+#' my_skim <- skim_with(numeric = list(median = median, mad = mad),
+#'                      append = FALSE)
+#' my_skim(faithful)
+#' 
+#' # If you want to remove a particular skimmer, set it to NULL
+#' # This removes the inline histogram
+#' my_skim <- skim_with(numeric = list(hist = NULL))
+#' my_skim(faithful)
+#' 
+#' # This works with multiple skimmers. Just match names to overwrite
+#' my_skim <- skim_with(numeric = list(iqr = IQR, p25 = NULL, p75 = NULL))
+#' my_skim(faithful)
+#' 
+#' # Alternatively, set `append = FALSE` to replace the skimmers of a type.
+#' my_skim <- skim_with(numeric = list(mean = mean, sd = sd), append = FALSE)
+#' 
+#' # Skimmers are unary functions. Partially apply arguments during assigment.
+#' # For example, you might want to remove NA values. Use `dplyr::funs()`
+#' # syntax for partial application.
+#' my_skim <- skim_with(numeric = list(iqr = IQR(., na.rm = TRUE)))
+#' 
+#' 
+#' # Set multiple types of skimmers simultaneously
+#' my_skim <- skim_with(numeric = list(mean = mean),
+#'                      character = list(len = length))
+#' 
+#' # Or pass the same as a list
+#' my_skimmers <- list(numeric = list(mean = mean),
+#'                     character = list(len = length))
+#' my_skim <- skim_with(!!!my_skimmers)
+#' 
+#' # Go back to defaults
+#' default_skim <- skim_with()
+#' idenctical(skim(faithful), default_skim(faithful))
+#' #> TRUE
+#' @export
+
+skim_with <- function(..., append = TRUE) {
+  validate_assignment(...)
+  local_skimmers <- purrr::map_if(list(...), ~!is.null(.x), assign_skimmers)
+  function(data, ...) {
+    .vars <- rlang::quos(...)
+    cols <- names(data)
+    if (length(.vars) == 0) {
+      selected <- cols
+    } else {
+      selected <- cols[tidyselect::vars_select(cols, !!! .vars)]
+    }
+    
+    grps <- dplyr::groups(data)
+    if (length(grps) > 0) {
+      group_variables <- selected %in% as.character(grps)
+      selected <- selected[!group_variables]
+    }
+    
+    variables <- data.frame(variable = selected, stringsAsFactors = FALSE)
+    nested <- dplyr::mutate(variables, purrr::map(variable, skim_one, data,
+                                                  local_skimmers, append))
[…]
