### 2018-05-27 e1d279f
Remove deprecated.
