+++ README.md
@@ -33,7 +33,8 @@ been incorporated in the master branch (and may not be):
 
     devtools::install_github("ropenscilabs/skimr", ref = "develop")
 
-Do not rely on APIs from the develop branch.
+Do not rely on APIs from the develop branch, as they are likely to
+change.
 
 Skim statistics in the console
 ------------------------------
@@ -54,12 +55,16 @@ Skim statistics in the console
     ## Skim summary statistics
     ##  n obs: 71 
     ##  n variables: 2 
-    ## ── Variable type:factor ────────────────────────────────────────────
-    ##  variable missing complete  n n_unique                         top_counts ordered
-    ##      feed       0       71 71        6 soy: 14, cas: 12, lin: 12, sun: 12   FALSE
-    ## ── Variable type:numeric ───────────────────────────────────────────
-    ##  variable missing complete  n   mean    sd  p0   p25 p50   p75 p100     hist
-    ##    weight       0       71 71 261.31 78.07 108 204.5 258 323.5  423 ▃▅▅▇▃▇▂▂
+    ## ── Variable type: double ─────────────────────────────────────────────────────────────
+    ## # A tibble: 1 x 12
+    ##   variable missing complete     n  mean    sd    p0   p25   p50   p75  p100 hist 
+    ## * <chr>      <int>    <int> <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <chr>
+    ## 1 weight         0       71    71  261.  78.1  108.  204.  258.  324.  423. ▆▆▇▇▃
+    ## ── Variable type: factor ─────────────────────────────────────────────────────────────
+    ## # A tibble: 1 x 7
+    ##   variable missing complete ordered     n n_unique top_counts                        
+    ## * <chr>      <int>    <int> <lgl>   <int>    <int> <chr>                             
+    ## 1 feed           0       71 FALSE      71        6 soy: 14, cas: 12, lin: 12, sun: 12
 
 ### Presentation is in a compact horizontal format:
 
@@ -68,15 +73,19 @@ Skim statistics in the console
     ## Skim summary statistics
     ##  n obs: 150 
     ##  n variables: 5 
-    ## ── Variable type:factor ────────────────────────────────────────────
-    ##  variable missing complete   n n_unique                       top_counts ordered
-    ##   Species       0      150 150        3 set: 50, ver: 50, vir: 50, NA: 0   FALSE
-    ## ── Variable type:numeric ───────────────────────────────────────────
-    ##      variable missing complete   n mean   sd  p0 p25  p50 p75 p100     hist
-    ##  Petal.Length       0      150 150 3.76 1.77 1   1.6 4.35 5.1  6.9 ▇▁▁▂▅▅▃▁
-    ##   Petal.Width       0      150 150 1.2  0.76 0.1 0.3 1.3  1.8  2.5 ▇▁▁▅▃▃▂▂
-    ##  Sepal.Length       0      150 150 5.84 0.83 4.3 5.1 5.8  6.4  7.9 ▂▇▅▇▆▅▂▂
-    ##   Sepal.Width       0      150 150 3.06 0.44 2   2.8 3    3.3  4.4 ▁▂▅▇▃▂▁▁
+    ## ── Variable type: double ─────────────────────────────────────────────────────────────
+    ## # A tibble: 4 x 12
+    ##   variable     missing complete     n  mean    sd    p0   p25   p50   p75  p100 hist 
+    ## * <chr>          <int>    <int> <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <chr>
[…]
