+++ README.Rmd
@@ -5,11 +5,7 @@ output:
 <!-- README.md is generated from README.Rmd. Please edit that file -->
 # skimr
 ```{r set-options, echo=FALSE, message=FALSE}
-library(pander)
-library(knitr)
 library(skimr)
-library(magrittr)
-skim_with_defaults()
 options(tibble.width = Inf)
 options(width = 120)
 ```
@@ -30,20 +26,22 @@ The current released version of `skimr` can be installed from CRAN. If you wish
 to install the current build of the next release you can do so using the
 following:
 
-``` r
+```{r, eval = FALSE}
 # install.packages("devtools")
 devtools::install_github("ropenscilabs/skimr")
 ```
+
 The APIs for this branch should be considered reasonably stable but still
 subject to change if an issue is discovered.
 
 To install the version with the most recent changes that have not yet been
 incorporated in the master branch (and may not be):
 
-``` r
+```{r, eval = FALSE}
 devtools::install_github("ropenscilabs/skimr", ref = "develop")
 ```
-Do not rely on APIs from the develop branch.
+
+Do not rely on APIs from the develop branch, as they are likely to change.
 
 ## Skim statistics in the console
 
@@ -58,92 +56,50 @@ complete, n, and sd.
 
 ### Separates variables by class:  
 
-```{r}
+```{r, render = knitr::normal_print}
 skim(chickwts)
 ```
 
 ### Presentation is in a compact horizontal format: 
-```{r}
+
+```{r, render = knitr::normal_print}
 skim(iris)
 ```
+
 ### Built in support for strings, lists and other column classes
 
-```{r}
+```{r, render = knitr::normal_print}
 skim(dplyr::starwars)
 ```
 
 ### Has a useful summary function
-```{r}
-skim(iris) %>% summary()
-```
 
+```{r, render = knitr::normal_print}
+skim(iris) %>%
+  summary()
+```
 
 ### Individual columns can be selected using tidyverse-style selectors
 
-```{r}
+```{r, render = knitr::normal_print}
 skim(iris, Sepal.Length, Petal.Length)
 ```
 
 ### Handles grouped data 
 
-`skim()` can handle data that has been grouped using `dplyr::group_by`.
+`skim()` can handle data that has been grouped using `dplyr::group_by()`.
 
-```{r}
-iris %>% dplyr::group_by(Species) %>% skim()
+```{r, render = knitr::normal_print}
+iris %>%
+  dplyr::group_by(Species) %>%
+  skim()
 ```
 
 ## Knitted results
 
 Simply skimming a data frame will produce the horizontal print
-layout shown above. When knitting you can also used enhanced rendering
-with kable and pander implementations. 
-
-### Options for kable and pander
-
-Enhanced print options are available by piping to `kable()` or `pander()`.
-These build on the [pander package](https://CRAN.R-project.org/package=pander)
-and the kable function of the [knitr package](https://CRAN.R-project.org/package=knitr)
-These examples show how the enhanced options should appear after 
-knitting, however your results may differ (see vignettes for details).
-
-
-### Option for kable. 
-Note that the results='asis' chunk option is used and  the
-`skimr::` namespace is used to prevent it being
-replaced by knitr::kable (which will result in the long skim_df 
-object being printed.)
-```{r results='asis'}
-skim(iris) %>% skimr::kable()
-```
-
-
-### Options for pander
-
-At times you may need `panderOptions('knitr.auto.asis', FALSE)`.
-
-```{r}
-skim(iris) %>% pander()
-```
-
-
-## `skim_df` object (long format)
-
-By default `skim()` prints beautifully in the console, but it also produces a
[…]
