### 2018-05-27 1a05016
Add functions for reshaping a skim_df
