### 2018-05-27 77dfca1
Update print methods.
