+++ R/skim_print.R
@@ -1,46 +1,47 @@
 #' Print `skim` objects
 #' 
+#' `skimr` has custom print methods for all supported objects. Default printing
+#' methods for `knitr`/ `rmarkdown` documents is also provided.
+#' 
 #' @param x Either a `skim_df`, `skim_vector` or `skim_summary` object.
 #' @param ... Further arguments passed to or from other methods.
 #' @name print
 NULL
 
-
-#' @describeIn print Prints a skimmed data frame (`skim_df` from [`skim()`])
+#' @describeIn print Print a skimmed data frame (`skim_df` from [`skim()`]).
 #' @export
-
-print.skim_df <- function(x, ...) {
-  defaults <- options(dplyr.show_progress = FALSE)
-  on.exit(options(defaults)
-  )
-  cat("Skim summary statistics\n")
-  cat(" n obs:", attr(x, "data_rows"), "\n")
-  cat(" n variables:", attr(x, "data_cols"), "\n")
-
-  grps <- dplyr::groups(x) 
-  if (!is.null(grps)) {
-    flat <- paste(grps, collapse = ", ")
-    cat(" group variables:", flat, "\n")
+print.skim_df <- function(x, include_summary = TRUE, ...) {
+  if (include_summary) {
+    cat("Skim summary statistics\n")
+    cat(" n obs:", attr(x, "data_rows"), "\n")
+    cat(" n variables:", attr(x, "data_cols"), "\n")
   }
-  
-  grouped <- dplyr::group_by(x, !!rlang::sym("type"))
-  dplyr::do(grouped, skim_render(., grps, print_impl, ...))
-  invisible(x)
+
+  by_type <- partition(x)
+  purrr::imap(by_type, print)
+  invisible(NULL)
 }
 
-
-#' @describeIn print Manages print for `skim_vector` objects.
+#' @describeIn print Print an entry within a partitioned `skim_df`.
 #' @export
+print.one_skim_df <- function(x, ..) {
+  variable_type <- paste("Variable type:", attr(x, "type"))
+  with_line <- cli::rule(line = 1, left = variable_type)
+  print(with_line)
+  NextMethod("print")
+}
 
-print.skim_vector <- function(x, ...) {
-  cat("\nSkim summary statistics\n")
-  skim_render(x, groups = as.null(), print_impl, ...)
+#' @describeIn print Print a `skim_list`, a list of `skim_df` objects.
+#' @export
+print.skim_list <- function(x, ..) {
+  nms <- names(x)
+  attributes(x) <- NULL
+  print(purrr::set_names(x, nms))
 }
 
 
 #' @describeIn print Print method for a `summary_skim_df` object.
 #' @export
- 
 print.summary_skim_df <- function(x, ...) {
   n_rows <- paste0("Number of Rows: ", x$n_rows, "   \n")
   n_cols <- paste0("Number of Columns: ", x$n_cols, "    \n")
@@ -61,60 +62,86 @@ print.summary_skim_df <- function(x, ...) {
       ,sep = "")
 }
 
-#' Print expanded skim tables with a simple caption
-#' @keywords internal
-#' @noRd
+#' Provide a default printing method for knitr.
+#' 
+#' Instead of standard R output, `knitr` and `RMarkdown` documents will have
+#' formatted [knitr::kable()] output on return. You can disable this by setting
+#' the chunk option `render = normal_print`.
+#' 
+#' The summary statistics for the original data frame can be disabled by setting
+#' the `knitr` chunk option `skimr_include_summary = FALSE`. See
+#' [knitr::opts_chunk] for more information. You can change the number of digits
+#' shown in the printed table with the `skimr_digits` chunk option.
[…]
