### 2018-05-27 19c9d89
Implement skim function lists.
