+++ R/skim.R
@@ -1,5 +1,3 @@
-globalVariables(".")
-
 #' Set or add the summary functions for a particular type of data
 #' 
 #' While skim is designed around having an opinionated set of defaults, you
@@ -13,61 +11,59 @@ globalVariables(".")
 #' also means that you need to assign the return of `skim_with()` before
 #' you can use it.
 #' 
-#' @param ... A list of functions, with an argument name that matches a
-#'  particular data type.
+#' You assign values within `skim_with` by using the [sfl()] helper (`skimr`
+#' function list). This helper behaves mostly like [dplyr::funs()], but lets
+#' you also identify which skimming functions you want to remove, by setting
+#' them to `NULL`.
+#' 
+#' @param ... One or more `skimmer_function_list` objects, with an argument name
+#'  that matches a particular data type.
 #' @param append Whether the provided options should be in addition to the
 #'  defaults already in `skim`. Default is `TRUE`.
 #' @return A new `skim()` function. This is callable. See [skim()] for
 #'  more details.
 #' @examples
-#' # Use new functions for numeric functions
-#' my_skim <- skim_with(numeric = list(median = median, mad = mad),
-#'                      append = FALSE)
+#' # Use new functions for numeric functions. If you don't provide a name,
+#' # one will be automatically generated.
+#' my_skim <- skim_with(numeric = sfl(median, mad), append = FALSE)
 #' my_skim(faithful)
 #' 
 #' # If you want to remove a particular skimmer, set it to NULL
 #' # This removes the inline histogram
-#' my_skim <- skim_with(numeric = list(hist = NULL))
+#' my_skim <- skim_with(numeric = sfl(hist = NULL))
 #' my_skim(faithful)
 #' 
 #' # This works with multiple skimmers. Just match names to overwrite
-#' my_skim <- skim_with(numeric = list(iqr = IQR, p25 = NULL, p75 = NULL))
+#' my_skim <- skim_with(numeric = sfl(iqr = IQR, p25 = NULL, p75 = NULL))
 #' my_skim(faithful)
 #' 
 #' # Alternatively, set `append = FALSE` to replace the skimmers of a type.
-#' my_skim <- skim_with(numeric = list(mean = mean, sd = sd), append = FALSE)
+#' my_skim <- skim_with(numeric = sfl(mean = mean, sd = sd), append = FALSE)
 #' 
 #' # Skimmers are unary functions. Partially apply arguments during assigment.
 #' # For example, you might want to remove NA values. Use `dplyr::funs()`
 #' # syntax for partial application.
-#' my_skim <- skim_with(numeric = list(iqr = IQR(., na.rm = TRUE)))
+#' my_skim <- skim_with(numeric = sfl(iqr = IQR(., na.rm = TRUE)))
 #' 
+#' # Or, use the `.args` argument from `dplyr::funs()`
+#' my_skim <- skim_with(numeric = sfl(median, mad, .args = list(na.rm = FALSE)))
 #' 
-#' # Set multiple types of skimmers simultaneously
-#' my_skim <- skim_with(numeric = list(mean = mean),
-#'                      character = list(len = length))
+#' # Set multiple types of skimmers simultaneously.
+#' skim2 <- skim_with(numeric = sfl(mean), character = sfl(length))
 #' 
 #' # Or pass the same as a list
-#' my_skimmers <- list(numeric = list(mean = mean),
-#'                     character = list(len = length))
+#' my_skimmers <- list(numeric = sfl(mean), character = sfl(length))
 #' my_skim <- skim_with(!!!my_skimmers)
-#' 
-#' # Go back to defaults
-#' default_skim <- skim_with()
-#' idenctical(skim(faithful), default_skim(faithful))
-#' #> TRUE
[…]
