+++ man/skim_with.Rd
@@ -2,25 +2,23 @@
 % Please edit documentation in R/skim.R
 \name{skim_with}
 \alias{skim_with}
+\alias{get_default_skimmers}
 \title{Set or add the summary functions for a particular type of data}
 \usage{
 skim_with(..., append = TRUE)
+
+get_default_skimmers()
 }
 \arguments{
-\item{...}{A list of functions, with an argument name that matches a
-particular data type.}
+\item{...}{One or more \code{skimmer_function_list} objects, with an argument name
+that matches a particular data type.}
 
 \item{append}{Whether the provided options should be in addition to the
 defaults already in \code{skim}. Default is \code{TRUE}.}
-
-\item{drop_new}{Whether types outside of the defaults should be discarded.}
 }
 \value{
-When setting values, \code{invisible(NULL)}. When looking up values a
-list. The names of the list match the classes that have assigned
-summary functions. With \code{\link[=show_skimmers]{show_skimmers()}}, each entry in the list is a
-character vector of function names. With \code{\link[=get_skimmers]{get_skimmers()}}, each entry
-in the list is itself a list of named functions.
+A new \code{skim()} function. This is callable. See \code{\link[=skim]{skim()}} for
+more details.
 }
 \description{
 While skim is designed around having an opinionated set of defaults, you
@@ -30,52 +28,51 @@ along with a list of named functions that you want to use instead of the
 defaults.
 }
 \details{
-This function is not pure. It sets values in within the package environment.
-This is an intentional design choice, with effects similar to setting
-options in base R. By setting options here for your entire session, you
-can continue to summarize using skim on its own.
+\code{skim_with()} is a closure: a function that returns a new function. This
+lets you have several skimming functions in a single R session, but it
+also means that you need to assign the return of \code{skim_with()} before
+you can use it.
+
+You assign values within \code{skim_with} by using the \code{\link[=sfl]{sfl()}} helper (\code{skimr}
+function list). This helper behaves mostly like \code{\link[dplyr:funs]{dplyr::funs()}}, but lets
+you also identify which skimming functions you want to remove, by setting
+them to \code{NULL}.
 }
+\section{Functions}{
+\itemize{
+\item \code{get_default_skimmers}: Get a list of classes with default skim functions
+}}
+
 \examples{
-# Use new functions for numeric functions
-skim_with(numeric = list(median = median, mad = mad), append = FALSE)
-skim(faithful)
+# Use new functions for numeric functions. If you don't provide a name,
+# one will be automatically generated.
+my_skim <- skim_with(numeric = sfl(median, mad), append = FALSE)
+my_skim(faithful)
 
 # If you want to remove a particular skimmer, set it to NULL
 # This removes the inline histogram
-skim_with(numeric = list(hist = NULL))
-skim(faithful)
+my_skim <- skim_with(numeric = sfl(hist = NULL))
+my_skim(faithful)
 
 # This works with multiple skimmers. Just match names to overwrite
-skim_with(numeric = list(iqr = IQR, p25 = NULL, p75 = NULL))
-skim(faithful)
+my_skim <- skim_with(numeric = sfl(iqr = IQR, p25 = NULL, p75 = NULL))
+my_skim(faithful)
 
 # Alternatively, set `append = FALSE` to replace the skimmers of a type.
-skim_with(numeric = list(mean = mean, sd = sd), append = FALSE)
+my_skim <- skim_with(numeric = sfl(mean = mean, sd = sd), append = FALSE)
 
 # Skimmers are unary functions. Partially apply arguments during assigment.
[…]
