+++ man/sfl.Rd
@@ -1,0 +1,46 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/skim.R
+\name{sfl}
+\alias{sfl}
+\title{Create a skimr function list}
+\usage{
+sfl(..., .args = list())
+}
+\arguments{
+\item{...}{A list of functions specified by:
+\itemize{
+\item Their name, \code{"mean"}
+\item The function itself, \code{mean}
+\item A call to the function with \code{.} as a dummy argument,
+\code{mean(., na.rm = TRUE)}
+}
+
+These arguments are automatically \link[rlang:quo]{quoted}. They
+support \link[rlang:quasiquotation]{unquoting} and splicing. See
+\code{vignette("programming")} for an introduction to these concepts.
+
+The following notations are \strong{not} supported, see examples:
+\itemize{
+\item An anonymous function, \code{function(x) mean(x, na.rm = TRUE)}
+\item An anonymous function in \pkg{purrr} notation, \code{~mean(., na.rm = TRUE)}
+}}
+
+\item{.args}{A named list of additional arguments to be added
+to all function calls.}
+}
+\value{
+A \code{skimr_function_list}, which contains a list of \code{fun_calls},
+returned by \code{\link[dplyr:funs]{dplyr::funs()}} and a list of skimming functions to drop.
+}
+\description{
+This is an extension of \code{\link[dplyr:funs]{dplyr::funs()}}. It is used to create a named list
+of functions. It also you also pass \code{NULL} to identify a skimming function
+that you wish to remove.
+}
+\details{
+Only functions that return a single value, working with \code{\link[dplyr:summarize]{dplyr::summarize()}},
+can be used within \code{sfl}.
+}
+\seealso{
+\code{\link[dplyr:funs]{dplyr::funs()}} and \code{\link[=skim_with]{skim_with()}}
+}

[…]
