+++ NAMESPACE
@@ -9,25 +9,28 @@ S3method(get_skimmers,default)
 S3method(get_skimmers,double)
 S3method(get_skimmers,factor)
 S3method(get_skimmers,integer)
+S3method(get_skimmers,list)
 S3method(get_skimmers,logical)
 S3method(get_skimmers,ts)
-S3method(kable,data.frame)
-S3method(kable,skim_df)
-S3method(kable,summary_skim_df)
+S3method(knit_print,one_skim_df)
+S3method(knit_print,skim_df)
+S3method(knit_print,skim_list)
+S3method(knit_print,summary_skim_df)
+S3method(print,one_skim_df)
 S3method(print,skim_df)
-S3method(print,skim_vector)
-S3method(print,spark)
+S3method(print,skim_list)
 S3method(print,summary_skim_df)
 S3method(summary,skim_df)
 export("%>%")
+export(bind)
 export(contains)
 export(ends_with)
 export(everything)
 export(fix_windows_histograms)
+export(get_default_skimmers)
 export(get_skimmers)
 export(inline_hist)
 export(inline_linegraph)
-export(kable)
 export(list_lengths_max)
 export(list_lengths_median)
 export(list_lengths_min)
@@ -43,22 +46,20 @@ export(n_missing)
 export(n_unique)
 export(num_range)
 export(one_of)
-export(show_formats)
+export(partition)
+export(sfl)
 export(skim)
-export(skim.data.frame)
-export(skim.default)
-export(skim.grouped_df)
-export(skim_format)
-export(skim_format_defaults)
 export(skim_tee)
-export(skim_to_list)
-export(skim_to_wide)
 export(skim_with)
 export(sorted_count)
 export(starts_with)
+export(top_counts)
 export(ts_end)
 export(ts_start)
+export(yank)
+importFrom(knitr,knit_print)
 importFrom(magrittr,"%>%")
+importFrom(rlang,"%||%")
 importFrom(tidyselect,contains)
 importFrom(tidyselect,ends_with)
 importFrom(tidyselect,everything)
