### 2018-05-28 ca43da3
Readme's updated for sfl interface.
