+++ R/skim_with.R
@@ -1,0 +1,200 @@
+#' Set or add the summary functions for a particular type of data
+#' 
+#' While skim is designed around having an opinionated set of defaults, you
+#' can use these functions to change the summary statistics that it returns.
+#' To do that, provide type you wish to change as an argument to this function,
+#' along with a list of named functions that you want to use instead of the
+#' defaults. 
+#' 
+#' `skim_with()` is a closure: a function that returns a new function. This
+#' lets you have several skimming functions in a single R session, but it
+#' also means that you need to assign the return of `skim_with()` before
+#' you can use it.
+#' 
+#' You assign values within `skim_with` by using the [sfl()] helper (`skimr`
+#' function list). This helper behaves mostly like [dplyr::funs()], but lets
+#' you also identify which skimming functions you want to remove, by setting
+#' them to `NULL`.
+#' 
+#' When `append = TRUE` and local skimmers have names matching the names of
+#' entries in the default `skim_function_list`, the values in the default list
+#' are overwritten. Similarly, if `NULL` values are passed within `sfl()`, these
+#' default skimmers are dropped. Otherwise, if `append = FALSE`, only the
+#' locally-provided skimming functions are used.
+#' 
+#' @param ... One or more `skimmer_function_list` objects, with an argument name
+#'  that matches a particular data type.
+#' @param append Whether the provided options should be in addition to the
+#'  defaults already in `skim`. Default is `TRUE`.
+#' @return A new `skim()` function. This is callable. See [skim()] for
+#'  more details.
+#' @examples
+#' # Use new functions for numeric functions. If you don't provide a name,
+#' # one will be automatically generated.
+#' my_skim <- skim_with(numeric = sfl(median, mad), append = FALSE)
+#' my_skim(faithful)
+#' 
+#' # If you want to remove a particular skimmer, set it to NULL
+#' # This removes the inline histogram
+#' my_skim <- skim_with(numeric = sfl(hist = NULL))
+#' my_skim(faithful)
+#' 
+#' # This works with multiple skimmers. Just match names to overwrite
+#' my_skim <- skim_with(numeric = sfl(iqr = IQR, p25 = NULL, p75 = NULL))
+#' my_skim(faithful)
+#' 
+#' # Alternatively, set `append = FALSE` to replace the skimmers of a type.
+#' my_skim <- skim_with(numeric = sfl(mean = mean, sd = sd), append = FALSE)
+#' 
+#' # Skimmers are unary functions. Partially apply arguments during assigment.
+#' # For example, you might want to remove NA values. Use `dplyr::funs()`
+#' # syntax for partial application.
+#' my_skim <- skim_with(numeric = sfl(iqr = IQR(., na.rm = TRUE)))
+#' 
+#' # Or, use the `.args` argument from `dplyr::funs()`
+#' my_skim <- skim_with(numeric = sfl(median, mad, .args = list(na.rm = FALSE)))
+#' 
+#' # Set multiple types of skimmers simultaneously.
+#' my_skim <- skim_with(numeric = sfl(mean), character = sfl(length))
+#' 
+#' # Or pass the same as a list
+#' my_skimmers <- list(numeric = sfl(mean), character = sfl(length))
+#' my_skim <- skim_with(!!!my_skimmers)
+#' @export
+skim_with <- function(..., append = TRUE) {
+  local_skimmers <- validate_assignment(...)
+  function(data, ...) {
+    .vars <- rlang::quos(...)
+    cols <- names(data)
+    if (length(.vars) == 0) {
+      selected <- cols
+    } else {
+      selected <- tidyselect::vars_select(cols, !!!.vars)
+    }
+    
+    grps <- dplyr::groups(data)
[…]
