+++ README.md
@@ -55,16 +55,16 @@ Skim statistics in the console
     ## Skim summary statistics
     ##  n obs: 71 
     ##  n variables: 2 
-    ## ── Variable type: double ─────────────────────────────────────────────────────────────
+    ## ── Variable type: factor ────────────────────────────────────────────────────────────────────────
+    ## # A tibble: 1 x 7
+    ##   variable missing complete     n ordered n_unique top_counts                        
+    ## * <chr>      <int>    <int> <int> <lgl>      <int> <chr>                             
+    ## 1 feed           0       71    71 FALSE          6 soy: 14, cas: 12, lin: 12, sun: 12
+    ## ── Variable type: numeric ───────────────────────────────────────────────────────────────────────
     ## # A tibble: 1 x 12
     ##   variable missing complete     n  mean    sd    p0   p25   p50   p75  p100 hist 
     ## * <chr>      <int>    <int> <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <chr>
     ## 1 weight         0       71    71  261.  78.1  108.  204.  258.  324.  423. ▆▆▇▇▃
-    ## ── Variable type: factor ─────────────────────────────────────────────────────────────
-    ## # A tibble: 1 x 7
-    ##   variable missing complete ordered     n n_unique top_counts                        
-    ## * <chr>      <int>    <int> <lgl>   <int>    <int> <chr>                             
-    ## 1 feed           0       71 FALSE      71        6 soy: 14, cas: 12, lin: 12, sun: 12
 
 ### Presentation is in a compact horizontal format:
 
@@ -73,7 +73,12 @@ Skim statistics in the console
     ## Skim summary statistics
     ##  n obs: 150 
     ##  n variables: 5 
-    ## ── Variable type: double ─────────────────────────────────────────────────────────────
+    ## ── Variable type: factor ────────────────────────────────────────────────────────────────────────
+    ## # A tibble: 1 x 7
+    ##   variable missing complete     n ordered n_unique top_counts                      
+    ## * <chr>      <int>    <int> <int> <lgl>      <int> <chr>                           
+    ## 1 Species        0      150   150 FALSE          3 set: 50, ver: 50, vir: 50, NA: 0
+    ## ── Variable type: numeric ───────────────────────────────────────────────────────────────────────
     ## # A tibble: 4 x 12
     ##   variable     missing complete     n  mean    sd    p0   p25   p50   p75  p100 hist 
     ## * <chr>          <int>    <int> <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <chr>
@@ -81,11 +86,6 @@ Skim statistics in the console
     ## 2 Sepal.Width        0      150   150  3.06 0.436 2.00  2.80   3.00  3.30  4.40 ▁▆▇▂▁
     ## 3 Petal.Length       0      150   150  3.76 1.77  1.00  1.60   4.35  5.10  6.90 ▇▁▆▇▂
     ## 4 Petal.Width        0      150   150  1.20 0.762 0.100 0.300  1.30  1.80  2.50 ▇▁▇▅▃
-    ## ── Variable type: factor ─────────────────────────────────────────────────────────────
-    ## # A tibble: 1 x 7
-    ##   variable missing complete ordered     n n_unique top_counts                      
-    ## * <chr>      <int>    <int> <lgl>   <int>    <int> <chr>                           
-    ## 1 Species        0      150 FALSE     150        3 set: 50, ver: 50, vir: 50, NA: 0
 
 ### Built in support for strings, lists and other column classes
 
[…]
