### 2018-05-28 fb89d56
Extend skimr using `sfl` calls within `get_skimmers`.
