+++ R/get_skimmers.R
@@ -1,78 +1,57 @@
-#' @include stats.R 
+#' @include skim_with.R stats.R 
 NULL
 
 #' Retrieve the summary functions for a specific data type
 #' 
 #' These functions are used to set the default skimming functions for a data
-#' type.
+#' type.  When creating your own set of skimming functions, call [sfl()]
+#' within a [get_skimmers()] method for your particular type. Your call to
+#' [sfl()] should also provide a matching class in the `.class` argument.
+#' Otherwise, it will not be possible to dynamically reassign your default
+#' functions when working interactively.
 #' 
-#' When creating your own set of skimming functions, call [make_skimmers()]
-#' within a [get_skimmers()].
+#' Summary functions are provided for the following classes:
+#' 
+#'  - `numeric`
+#'  - `integer`
+#'  - `character`
+#'  - `factor`
+#'  - `logical`
+#'  - `complex`
+#'  - `Date`
+#'  - `POSIXct`
+#'  - `ts`
+#'  - `AsIs`
+#'
+#' Call [get_default_skimmers()] to see the functions for each.
 #' 
 #' @param column An atomic vector or list. A column from a data frame.
-#' @param local_skimmers An environment. Used by [skim()].
-#' @param append A logical scalar. Used by [skim()]. Whether new skim functions
-#'   should be added to the existing set of functions or whether they new
-#'   functions should replace the old.
-#' @inheritParams dplyr::funs
-#' @return A list with two entries:
-#' 
-#'   - `funs` is a `fun_list`. The same as [dplyr::funs()].
-#'   - `type` is the type of skimmer called.
-#'
-#' @seealso [dplyr::funs()]
+#' @return A `skim_function_list` object.
+#' @seealso [sfl()]
+#' @examples
+#' # Defining default skimming functions for a new class, `my_class`.
+#' # Note that the class argument is required for dynamic reassignment.
+#' get_skimmers.my_class <- function(colum) {
+#'   sfl(
+#'     .class = "my_class",
+#'     mean,
+#'     sd)
+#' }
 #' @export
-get_skimmers <- function(column, local_skimmers = new.env(), append = TRUE) {
+get_skimmers <- function(column) {
   UseMethod("get_skimmers")
 }
 
-#' @rdname get_skimmers
 #' @export
-make_skimmer <- function(type, local_skimmers, append, ...) {
-  if (append) {
-    locals <- local_skimmers[[type]]
-    list(funs = merge_skimmers(dplyr::funs(...), locals),
-         type = type)
-  } else {
-    list(funs = local_skimmers[[type]]$keep %||% dplyr::funs(...),
-         type = type)
-  }
-}
-
-merge_skimmers <- function(default, user_provided) {
-  if (length(user_provided$keep) > 0) {
-    default[names(user_provided$keep)] <- user_provided$keep
-  }
-  
-  if (length(user_provided$drop) > 0) {
-    default[user_provided$drop] <- NULL
-  }
-  
-  default
+get_skimmers.default <- function(column) {
+  fallback <- get_skimmers(character())
+  sfl(.class = "default", !!!fallback$keep)
 }
 
 #' @export
-get_skimmers.default <- function(column, local_skimmers = new.env(),
-                                 append = TRUE) {
-  all_matches <- local_skimmers[class(column)]
-  first_match <- purrr::detect(all_matches, ~!is.null(.x))
-  
-  if (is.null(first_match)) {
-    msg <- "Couldn't find skimmers for class: %s; Using `character`."
-    warning(sprintf(msg, class(column)), call. = FALSE)
-    get_skimmers(as.character(column), local_skimmers)
-  } else {
[…]
