+++ man/sfl.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/skim.R
+% Please edit documentation in R/sfl.R
 \name{sfl}
 \alias{sfl}
 \title{Create a skimr function list}
 \usage{
-sfl(..., .args = list())
+sfl(..., .args = list(), .class = "")
 }
 \arguments{
 \item{...}{A list of functions specified by:
@@ -27,6 +27,9 @@ The following notations are \strong{not} supported, see examples:
 
 \item{.args}{A named list of additional arguments to be added
 to all function calls.}
+
+\item{.class}{A character scalar. This is used to match locally-provided
+skimmers with defaults. See \code{\link[=get_skimmers]{get_skimmers()}} for more detail.}
 }
 \value{
 A \code{skimr_function_list}, which contains a list of \code{fun_calls},
@@ -35,12 +38,9 @@ returned by \code{\link[dplyr:funs]{dplyr::funs()}} and a list of skimming funct
 \description{
 This is an extension of \code{\link[dplyr:funs]{dplyr::funs()}}. It is used to create a named list
 of functions. It also you also pass \code{NULL} to identify a skimming function
-that you wish to remove.
-}
-\details{
-Only functions that return a single value, working with \code{\link[dplyr:summarize]{dplyr::summarize()}},
-can be used within \code{sfl}.
+that you wish to remove. Only functions that return a single value, working
+with \code{\link[dplyr:summarize]{dplyr::summarize()}}, can be used within \code{sfl}.
 }
 \seealso{
-\code{\link[dplyr:funs]{dplyr::funs()}} and \code{\link[=skim_with]{skim_with()}}
+\code{\link[dplyr:funs]{dplyr::funs()}}, \code{\link[=skim_with]{skim_with()}} and \code{\link[=get_skimmers]{get_skimmers()}}.
 }
