### 2018-05-28 eb82878
Remove "spark" class.
