### 2018-05-31 2699b88
Use the summarize-style top counts instead of sorted counts.
