### 2018-05-31 9f28a0b
Force use of a data.frame
