### 2018-05-31 d295769
Stop if arguments are empty.
