### 2018-05-31 4895983
Avoid recursive definition.
