### 2018-06-05 35bdbb5
Fix wrong variable name in test, thanks Prof. Ripley!
