+++ tests/testthat/test-skim_print.R
@@ -25,6 +25,7 @@ test_that("Skim prints a special header for vectors", {
 test_that("Skim collapses counts and other multivalue stats into one cell", {
   input <- skim(iris)
   expect_output(print(skim, "set: 50, ver: 50, vir: 50, NA: 0"))
+  expect_output(print(input, "set: 50, ver: 50, vir: 50, NA: 0"))
 })
 
 test_that("Skim aligns numeric vectors at the decimal point by default", {
