### 2018-06-05 dfdf5a8
Add minimum version for tidy select.
