### 2018-06-05 362d1ef
Try adding bug level to tidy select version.
