### 2018-06-05 #175 @elinw
Okay here's an update @jimhester @jenniferthompson @michaelquinn32
As of now all changes have been merged into the master branch, and I'm planning to do a release of 1.0.3 because I got a message from Brian Ripley about a typo in a test (I can't believe that all the eyes that looked at this before missed it).
I think we have responded to all of the issues except the possibility of changing the API.  Michael has written a road map for skimr 2.0 that addresses that along with some architectural issues.  But I think over the course of the discussion here and in the skimr repository we all came to agreement that cosmetic changes  that make the fact that the skim print results do not represent a single data frame more clear would be okay for skimr 1.
