### 2018-06-06 9a28a81
Change .class to .type for more consistent arguments.
