+++ man/get_skimmers.Rd
@@ -19,7 +19,7 @@ A \code{skim_function_list} object.
 These functions are used to set the default skimming functions for a data
 type.  When creating your own set of skimming functions, call \code{\link[=sfl]{sfl()}}
 within a \code{\link[=get_skimmers]{get_skimmers()}} method for your particular type. Your call to
-\code{\link[=sfl]{sfl()}} should also provide a matching class in the \code{.class} argument.
+\code{\link[=sfl]{sfl()}} should also provide a matching class in the \code{.type} argument.
 Otherwise, it will not be possible to dynamically reassign your default
 functions when working interactively.
 }
@@ -45,7 +45,7 @@ Call \code{\link[=get_default_skimmers]{get_default_skimmers()}} to see the func
 # Note that the class argument is required for dynamic reassignment.
 get_skimmers.my_class <- function(colum) {
   sfl(
-    .class = "my_class",
+    .type = "my_class",
     mean,
     sd)
 }
