### 2018-06-06 e414a79
Fix skimmer merging
