### 2018-06-07 #175 @maelle
Thanks for the updates @elinw and @michaelquinn32, very cool to see such active development?

@jimhester and @jenniferthompson , are you happy with the current version of `skimr`?
