### 2018-06-07 7e9ff9d
Standardize format of `skimr` & fix typo
