+++ README.Rmd
@@ -147,7 +147,7 @@ skim(mtcars) %>% dplyr::filter(stat=="hist")
 
 ## Customizing skimr
 
-Although skimr provides opinionated defaults, it is highly customizable.
+Although `skimr` provides opinionated defaults, it is highly customizable.
 Users can specify their own statistics, change the formatting of results,
 create statistics for new classes and develop skimmers for data structures
 that are not data frames.
@@ -174,9 +174,9 @@ skim_with_defaults()
 
 ### Change formatting
 
-Skimr provides a set of default formats that allow decimals in columns to
+`skimr` provides a set of default formats that allow decimals in columns to
 be aligned, a reasonable number of decimal places for numeric data, and a 
-representation of dates.  Users can view thes with `show_formats()` and
+representation of dates.  Users can view these with `show_formats()` and
 modify them with `skim_format()`.
 
 ### Skimming other objects
@@ -229,6 +229,6 @@ Microsoft Word and Libre Office Write.
 
 We welcome issue reports and pull requests, including potentially adding
 support for commonly used variable classes. However, in general, we encourage
-users to take advantage of skimr's flexibility to add their own customized
+users to take advantage of `skimr`'s flexibility to add their own customized
 classes. Please see the [contributing](CONTRIBUTING.md) and
 [conduct](CONDUCT.md) documents.
