### 2018-06-20 c7a2897
Throw a warning if a class doesn't have defaults.
