### 2018-06-20 e68b722
knit_print options should be NULL by default.
