### 2018-06-20 9f9450d
check for named values in skim_with
