+++ NAMESPACE
@@ -6,11 +6,12 @@ S3method(get_skimmers,POSIXct)
 S3method(get_skimmers,character)
 S3method(get_skimmers,complex)
 S3method(get_skimmers,default)
-S3method(get_skimmers,double)
+S3method(get_skimmers,difftime)
 S3method(get_skimmers,factor)
 S3method(get_skimmers,integer)
 S3method(get_skimmers,list)
 S3method(get_skimmers,logical)
+S3method(get_skimmers,numeric)
 S3method(get_skimmers,ts)
 S3method(knit_print,one_skim_df)
 S3method(knit_print,skim_df)
@@ -36,7 +37,6 @@ export(list_lengths_median)
 export(list_lengths_min)
 export(list_max_length)
 export(list_min_length)
-export(make_skimmer)
 export(matches)
 export(max_char)
 export(min_char)
