### 2018-06-20 cff6e1b
Add tests
