### 2018-06-20 #175 @jenniferthompson
Sorry for the delay, y'all (jet lag is very real) - just looked at the updated version and it looks super! :tada: Awesome work, @elinw @michaelquinn32.

Going to check out the version 2.0 roadmap now :running_woman:
