### 2018-06-21 7a0f43e
Export missing docs.
