### 2018-06-21 848d4be
Force naming from unnamed sfls. summarize_all() behaves differently in this case.
