### 2018-06-22 #175 @michaelquinn32
I really appreciate all of the hard work everyone has put in. `skimr` has improved in so many ways thanks to this process. Thanks!
