### 2018-06-23 19dbf29
Use the cite tag around package name in README
