+++ README.md
@@ -7,8 +7,8 @@ Status](https://travis-ci.org/ropenscilabs/skimr.svg?branch=master)](https://tra
 [![codecov](https://codecov.io/gh/ropenscilabs/skimr/branch/master/graph/badge.svg)](https://codecov.io/gh/ropenscilabs/skimr)
 [![](https://badges.ropensci.org/175_status.svg)](https://github.com/ropensci/onboarding/issues/175)
 
-`skimr` provides a frictionless approach to summary statistics which
-conforms to the [principle of least
+<cite>skimr</cite> provides a frictionless approach to summary
+statistics which conforms to the [principle of least
 surprise](https://en.wikipedia.org/wiki/Principle_of_least_astonishment),
 displaying summary statistics the user can skim quickly to understand
 their data. It handles different data types and returns a `skim_df`
@@ -18,9 +18,9 @@ human reader.
 Installation
 ------------
 
-The current released version of `skimr` can be installed from CRAN. If
-you wish to install the current build of the next release you can do so
-using the following:
+The current released version of <cite>skimr</cite> can be installed from
+CRAN. If you wish to install the current build of the next release you
+can do so using the following:
 
     # install.packages("devtools")
     devtools::install_github("ropenscilabs/skimr")
@@ -38,13 +38,13 @@ Do not rely on APIs from the develop branch.
 Skim statistics in the console
 ------------------------------
 
-`skimr`:
+<cite>skimr</cite>:
 
 -   Provides a larger set of statistics than `summary()`, including
     missing, complete, n, and sd.
 -   reports each data types separately
 -   handles dates, logicals, and a variety of other types
--   supports spark-bar and spark-line based on [Hadley Wickham's pillar
+-   supports spark-bar and spark-line based on [Hadley Wickham’s pillar
     package](https://github.com/hadley/pillar).
 
 ### Separates variables by class:
@@ -54,10 +54,12 @@ Skim statistics in the console
     ## Skim summary statistics
     ##  n obs: 71 
     ##  n variables: 2 
-    ## ── Variable type:factor ────────────────────────────────────────────
+    ## 
+    ## ── Variable type:factor ───────────────────────────────────────────────────────────────────────────────────────────────
     ##  variable missing complete  n n_unique                         top_counts ordered
     ##      feed       0       71 71        6 soy: 14, cas: 12, lin: 12, sun: 12   FALSE
-    ## ── Variable type:numeric ───────────────────────────────────────────
+    ## 
+    ## ── Variable type:numeric ──────────────────────────────────────────────────────────────────────────────────────────────
     ##  variable missing complete  n   mean    sd  p0   p25 p50   p75 p100     hist
     ##    weight       0       71 71 261.31 78.07 108 204.5 258 323.5  423 ▃▅▅▇▃▇▂▂
 
@@ -68,10 +70,12 @@ Skim statistics in the console
     ## Skim summary statistics
     ##  n obs: 150 
     ##  n variables: 5 
-    ## ── Variable type:factor ────────────────────────────────────────────
+    ## 
+    ## ── Variable type:factor ───────────────────────────────────────────────────────────────────────────────────────────────
     ##  variable missing complete   n n_unique                       top_counts ordered
     ##   Species       0      150 150        3 set: 50, ver: 50, vir: 50, NA: 0   FALSE
[…]
