### 2018-06-25 199a371
Fix missing namespace warnings, modify or comment in vignettes that is replaced in v2, regenerate help pages. There are still some warnings, however.
