+++ R/skim_print.R
@@ -29,7 +29,7 @@ print.skim_df <- function(x, include_summary = TRUE, ...) {
 
 #' @describeIn print Print an entry within a partitioned `skim_df`.
 #' @export
-print.one_skim_df <- function(x, ..) {
+print.one_skim_df <- function(x, ...) {
   variable_type <- paste("Variable type:", attr(x, "type"))
   with_line <- cli::rule(line = 1, left = variable_type)
   print(with_line)
@@ -38,7 +38,7 @@ print.one_skim_df <- function(x, ..) {
 
 #' @describeIn print Print a `skim_list`, a list of `skim_df` objects.
 #' @export
-print.skim_list <- function(x, ..) {
+print.skim_list <- function(x, ...) {
   nms <- names(x)
   attributes(x) <- NULL
   print(purrr::set_names(x, nms))
@@ -147,6 +147,6 @@ knit_print.summary_skim_df <- function(x, options = NULL, ...) {
   df_name <- paste0("Name: ", x[["df_name"]])
   
   kframe <- data.frame(df_name, n_rows, n_cols)
-  list(Summary = kable(kframe),
-       `Type counts` = kable(x$type_frequencies))
+  list(Summary = knitr::kable(kframe),
+       `Type counts` = knitr::kable(x$type_frequencies))
 }
