### 2018-06-25 7e2d7c7
Fix failing test and failing example.
