### 2018-06-25 1823129
Update some documentation (add some @params) and fix some typos.
