+++ man/skim.Rd
@@ -10,10 +10,14 @@ skim(data, ...)
 skim_tee(.data, ..., skim_fun = skim)
 }
 \arguments{
+\item{data}{The data frame that is skimmed and returned.}
+
 \item{...}{Columns to select for skimming. When none are provided, the
 default is to skim all columns.}
 
 \item{.data}{A tibble, or an object that can be coerced into a tibble.}
+
+\item{skim_fun}{The skim function used.}
 
 \item{skim}{The skimming function to use in \code{skim_tee()}.}
 }
@@ -38,9 +42,9 @@ If you just want to see the printed output, call \code{skim_tee()} instead.
 This function returns the original data. \code{skim_tee()} uses the default
 \code{skim()}, but you can replace it with the \code{skim} argument.
 
-If you want to work with a data frame that resembles the printed output,
-call \code{\link[=skim_to_wide]{skim_to_wide()}} or for a named list of data frames by type
-\code{\link[=skim_to_list]{skim_to_list()}}. Note that all of the columns in the data frames produced
+If you want to work with a data frame that is long (tidy),
+call \code{skim_to_long()} or for a named list of data frames by type
+\code{skim_to_list()}. Note that all of the columns in the data frames produced
 by these functions are character. The intent is that you will be processing
 the \strong{printed} result further, not the original data.
 
