### 2018-06-25 31361f0
Get rid of message about non existing linked document.
