### 2018-06-25 335330b
Fix error when using list in skim_with().
