### 2018-06-30 #175 @maelle
@elinw @michaelquinn32 :wave: could you soon tackle the list above? Thank you! 😃
