+++ README.md
@@ -3,8 +3,8 @@ skimr
 =====
 
 [![Build
-Status](https://travis-ci.org/ropenscilabs/skimr.svg?branch=master)](https://travis-ci.org/ropenscilabs/skimr)
-[![codecov](https://codecov.io/gh/ropenscilabs/skimr/branch/master/graph/badge.svg)](https://codecov.io/gh/ropenscilabs/skimr)
+Status](https://travis-ci.org/ropensci/skimr.svg?branch=master)](https://travis-ci.org/ropensci/skimr)
+[![codecov](https://codecov.io/gh/ropensci/skimr/branch/master/graph/badge.svg)](https://codecov.io/gh/ropensci/skimr)
 [![](https://badges.ropensci.org/175_status.svg)](https://github.com/ropensci/onboarding/issues/175)
 
 `skimr` provides a frictionless approach to summary statistics which
@@ -54,10 +54,12 @@ Skim statistics in the console
     ## Skim summary statistics
     ##  n obs: 71 
     ##  n variables: 2 
-    ## ── Variable type:factor ────────────────────────────────────────────
+    ## 
+    ## ── Variable type:factor ──────────────────────────────────────────────────────────────────────────────────────────────────
     ##  variable missing complete  n n_unique                         top_counts ordered
     ##      feed       0       71 71        6 soy: 14, cas: 12, lin: 12, sun: 12   FALSE
-    ## ── Variable type:numeric ───────────────────────────────────────────
+    ## 
+    ## ── Variable type:numeric ─────────────────────────────────────────────────────────────────────────────────────────────────
     ##  variable missing complete  n   mean    sd  p0   p25 p50   p75 p100     hist
     ##    weight       0       71 71 261.31 78.07 108 204.5 258 323.5  423 ▃▅▅▇▃▇▂▂
 
@@ -68,10 +70,12 @@ Skim statistics in the console
     ## Skim summary statistics
     ##  n obs: 150 
     ##  n variables: 5 
-    ## ── Variable type:factor ────────────────────────────────────────────
+    ## 
+    ## ── Variable type:factor ──────────────────────────────────────────────────────────────────────────────────────────────────
     ##  variable missing complete   n n_unique                       top_counts ordered
     ##   Species       0      150 150        3 set: 50, ver: 50, vir: 50, NA: 0   FALSE
-    ## ── Variable type:numeric ───────────────────────────────────────────
+    ## 
+    ## ── Variable type:numeric ─────────────────────────────────────────────────────────────────────────────────────────────────
     ##      variable missing complete   n mean   sd  p0 p25  p50 p75 p100     hist
     ##  Petal.Length       0      150 150 3.76 1.77 1   1.6 4.35 5.1  6.9 ▇▁▁▂▅▅▃▁
     ##   Petal.Width       0      150 150 1.2  0.76 0.1 0.3 1.3  1.8  2.5 ▇▁▁▅▃▃▂▂
@@ -85,7 +89,8 @@ Skim statistics in the console
     ## Skim summary statistics
     ##  n obs: 87 
     ##  n variables: 13 
-    ## ── Variable type:character ─────────────────────────────────────────
+    ## 
+    ## ── Variable type:character ───────────────────────────────────────────────────────────────────────────────────────────────
     ##    variable missing complete  n min max empty n_unique
     ##   eye_color       0       87 87   3  13     0       15
     ##      gender       3       84 87   4  13     0        4
@@ -94,15 +99,18 @@ Skim statistics in the console
     ##        name       0       87 87   3  21     0       87
     ##  skin_color       0       87 87   3  19     0       31
     ##     species       5       82 87   3  14     0       37
[…]
