+++ README.Rmd
@@ -14,8 +14,8 @@ options(tibble.width = Inf)
 options(width = 120)
 ```
 
-[![Build Status](https://travis-ci.org/ropenscilabs/skimr.svg?branch=master)](https://travis-ci.org/ropenscilabs/skimr)
-[![codecov](https://codecov.io/gh/ropenscilabs/skimr/branch/master/graph/badge.svg)](https://codecov.io/gh/ropenscilabs/skimr)
+[![Build Status](https://travis-ci.org/ropensci/skimr.svg?branch=master)](https://travis-ci.org/ropensci/skimr)
+[![codecov](https://codecov.io/gh/ropensci/skimr/branch/master/graph/badge.svg)](https://codecov.io/gh/ropensci/skimr)
 [![](https://badges.ropensci.org/175_status.svg)](https://github.com/ropensci/onboarding/issues/175)
 
 `skimr` provides a frictionless approach to summary statistics which conforms
@@ -232,3 +232,5 @@ support for commonly used variable classes. However, in general, we encourage
 users to take advantage of skimr's flexibility to add their own customized
 classes. Please see the [contributing](CONTRIBUTING.md) and
 [conduct](CONDUCT.md) documents.
+
+[![ropensci_footer](https://ropensci.org/public_images/ropensci_footer.png)](https://ropensci.org)
