+++ README.Rmd
@@ -233,4 +233,4 @@ users to take advantage of skimr's flexibility to add their own customized
 classes. Please see the [contributing](CONTRIBUTING.md) and
 [conduct](CONDUCT.md) documents.
 
-[![ropensci_footer](https://ropensci.org/public_images/ropensci_footer.png)](https://ropensci.org)
+[![ropenci_footer]](https://ropensci.org/public_images/ropensci_footer.png)](https://ropensci.org)]
