### 2018-06-30 #175 @elinw
- [x] Transfer the repo to the rOpenSci organization under "Settings" in your repo. I have invited you to a team that should allow you to do so. I'll make you both admins once you do.
- [ ]  Fix any links in badges for CI and coverage to point to the ropensci URL. We don't transfer Appveyor projects from personal accounts any more but since the project is under ropenscilabs we'll transfer it to ropensci account, once you've transferred the repo I'll activate the project.
  add rOpenSci footer to README
- [ ] [![ropensci_footer](https://ropensci.org/public_images/ropensci_footer.png)](https://ropensci.org)

About Appveyor that currently doesn't have a badge because of some problem we had before but we do run it off of @michaelquinn32's account.  So what to do?  We're happy to start from scratch on it.
