### 2018-06-30 #175 @maelle
You can keep the Appveyor project under @michaelquinn32's account if you prefer!
