+++ README.Rmd
@@ -16,7 +16,6 @@ options(width = 120)
 
 [![Build Status](https://travis-ci.org/ropensci/skimr.svg?branch=master)](https://travis-ci.org/ropensci/skimr)
 [![codecov](https://codecov.io/gh/ropensci/skimr/branch/master/graph/badge.svg)](https://codecov.io/gh/ropensci/skimr)
-[![](https://badges.ropensci.org/175_status.svg)](https://github.com/ropensci/onboarding/issues/175)
 
 `skimr` provides a frictionless approach to summary statistics which conforms
 to the [principle of least surprise](https://en.wikipedia.org/wiki/Principle_of_least_astonishment),
