### 2018-06-30 #175 @elinw
I think it makes sense to put it in ropensci.
