### 2018-07-01 ca6768e
Update skim.R
