+++ R/skim.R
@@ -15,11 +15,11 @@
 #' This function returns the original data. `skim_tee()` uses the default
 #' `skim()`, but you can replace it with the `skim` argument.
 #' 
-#' If you want to work with a data frame that is long (tidy),
-#' call `skim_to_long()` or for a named list of data frames by type
-#' `skim_to_list()`. Note that all of the columns in the data frames produced
-#' by these functions are character. The intent is that you will be processing
-#' the **printed** result further, not the original data.
+#' The data frame produced by `skim` is wide and parse, since not all summmary
+#' functions appear for all types. See [partition()] and [yank()] for methods
+#' for transforming this wide data frame. The first function splits it into
+#' a list, with each entry corresponding to a data type. The latter pulls
+#' a single subtable for a particular type from the `skim_df`.
 #' 
 #' `skim()` is designed to operate in pipes and to generally play nicely with
 #' other `tidyverse` functions. This means that you can use `tidyselect` helpers
