### 2018-07-01 044221b
Update skim documentation to use new reshaping API. Add more reshaping details.
