### 2018-07-01 #175 @maelle
Closing this, we can sort out CI details elsewhere and this way the peer-review badge will become green. 😉
