### 2018-07-18 #175 @stefaniebutland
@elinw Tired of writing about `skimr` yet? Or might you be interested in writing a post for the rOpenSci blog?
