### 2019-03-30 #175 @elinw
@jimhester  @jenniferthompson  @maelle  I think we are finally close to release for v2 which addresses most of the issues we discussed, specifically the format of the skim object.  We still need to rewrite vignettes and the readme.  We'll be adding a pkgdown site although we're not making that as a blocker for the release.  I think there is one important issue to resolve.  (We'll probably do one more v1 release since we have fixed a few issues there in the past month.)  @michaelquinn32  anything to add?
